==> 00_data_process_ple_react.Rmd <==
---
title: "01_data_process_ple_react"
author: "Levente Ronai"
output: html_document
---

loading packages

```{r}

options(scipen = 999)

suppressPackageStartupMessages(library(psych))
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(git2r))
suppressPackageStartupMessages(library(lavaan))
suppressPackageStartupMessages(library(semTools))

```

importing datasets

```{r}

big_data <- read.csv("data_processed/big_data.csv")

dat_cleaned <- 
  big_data %>% 
  select(id_esm, session_esm, created_esm, age, gender, edu, city, start_date, date_esm, date_diary, three_day_block,
         day, beep, event_pleasantness, sad, afraid, unrested, angry, irritated,
         contains(c("mss_", "mzq", "ple_", "stressor", "dprime")), -contains(c("minutes")))
  select(id_esm, session_esm, created_esm, age, gender, edu, city, start_date, date_esm, date_diary, ended_diary, three_day_block, day, beep, event_pleasantness, sad, afraid, unrested, angry, irritated, contains(c("mss_", "ple_", "stressor", "dprime")), -contains(c("minutes"))) %>% 
mutate(ended_diary = as_date(ended_diary))

write.csv(dat_cleaned, "data_processed/dat_cleaned.csv", row.names = F)

```

```{r}

dat_cleaned  <-  
  dat_cleaned %>%
  mutate(gender = factor(gender, levels = c(1, 2, 3), labels = c("Female", "Male", "Other")),
         event_unpleasantness = 8-event_pleasantness)

```


multilevel confirmatory factor analysis for psychotic-like experiences (within/between levels)

```{r, warning = T}

cfa_ple <-
  dat_cleaned %>%
  filter(!is.na(date_esm)) %>%
  select(id_esm, date_esm, beep,
         ple_control, ple_strange, ple_thoughts, ple_suspicion,
         ple_treatment, ple_perception, ple_thougtscontrol, ple_familiarstrange) %>%
  filter_at(vars(ple_control:ple_familiarstrange), all_vars(!is.na(.))) %>%
  distinct() %>%
  ungroup()

cfa_ple_syntax <- '

  level: 1

      ple_within   =~    ple_control + ple_strange + ple_thoughts + ple_suspicion +
                         ple_treatment + ple_perception + ple_thougtscontrol + ple_familiarstrange
                         
                         ple_perception	~~	ple_familiarstrange
                         ple_thoughts	~~	ple_familiarstrange


  level: 2

      ple_between   =~   ple_control + ple_strange + ple_thoughts + ple_suspicion +
                         ple_treatment + ple_perception + ple_thougtscontrol + ple_familiarstrange
                         
                        ple_control	~~	ple_strange

'

fit_ple <- cfa(model = cfa_ple_syntax, data = cfa_ple, cluster = "id_esm", std.lv = TRUE, estimator = "MLR")

```


```{r, warning = T}
fitMeasures(fit_ple, c("chisq", "df", "pvalue", "cfi", "tli", "rmsea", "srmr_within", "srmr_between"))
```


```{r, warning = T}

summary(fit_ple, fit.measures = TRUE, standardize = TRUE)


modificationindices(fit_ple) %>%
  select("lhs", "op", "rhs", "level", "mi") %>%
  filter(mi > 11) %>%
  arrange(desc(mi)) %>%
  head(n = 10) %>%
  print()

```

```{r, warning = T}

# within-person factor loadings
(inspect(fit_ple, what = "std")$within$lambda)

# between-person factor loadings
(inspect(fit_ple, what = "std")$id_esm$lambda)

# reliability of NA model
(semTools::reliability(fit_ple))

```


```{r, warning = T}

dat_ple_lev_2 <-
  cfa_ple %>%
  select(id_esm) %>%
  distinct() %>%
  cbind(lavPredict(fit_ple, newdata = cfa_ple[4:11], level = 2L))

dat_ple_lev_1 <-
  cfa_ple %>%
  cbind(lavPredict(fit_ple, newdata = cfa_ple[4:11], level = 1L)) %>%
  select(-c(ple_control:ple_familiarstrange))

dat_ple <-
  dat_ple_lev_1 %>%
  left_join(dat_ple_lev_2, by = "id_esm") %>%
  mutate(ple_cfa =  ple_within + ple_between)

dat_cleaned <-
  dat_cleaned %>% 
  left_join(dat_ple, by = c("id_esm", "date_esm", "beep"))


```

cronbach's alpha for mss (exclude #10 and #5 due to negative variances)

```{r}
# positive schizotypy: 2, 5, 8, 11, 14, 17, 20, 23, 26, 29, 32, 35, 38
mss_pos <- 
  dat_cleaned %>%
  select(id_esm, MSS_BH_8, MSS_BH_11, MSS_BH_14, 
           MSS_BH_23, MSS_BH_32, MSS_BH_20, MSS_BH_26, MSS_BH_29, MSS_BH_2, MSS_BH_38, MSS_BH_17, MSS_BH_35) %>% 
  # MSS_BH_5
  distinct() %>% 
  select(-id_esm)

# negative schizotypy: 1, 4R, 7, 10R, 13, 16, 19, 22, 25R, 28, 31, 34, 37R
mss_neg <- 
  dat_cleaned %>%
  select(id_esm, MSS_BH_1, MSS_BH_28, MSS_BH_13, MSS_BH_34, MSS_BH_19, MSS_BH_4, MSS_BH_7, MSS_BH_16,
             MSS_BH_25, MSS_BH_31, MSS_BH_37, MSS_BH_22) %>% 
  # MSS_BH_10
  distinct() %>% 
  select(-id_esm)

# disorganized schizotypy: 3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 33, 36  
mss_dis <- 
  dat_cleaned %>%
  select(id_esm, MSS_BH_3, MSS_BH_6, MSS_BH_9, 
           MSS_BH_12, MSS_BH_15, MSS_BH_18, 
           MSS_BH_21, MSS_BH_24, MSS_BH_27, 
           MSS_BH_30, MSS_BH_36, MSS_BH_33) %>%
  distinct() %>% 
  select(-id_esm)
```


```{r}
psych::alpha(mss_pos, check.keys=TRUE)
```

```{r}
psych::alpha(mss_neg, check.keys=TRUE)
```


```{r}
psych::alpha(mss_dis, check.keys=TRUE)
```

# CFA for MSS

```{r}

# positive schizotypy: 2, 5, 8, 11, 14, 17, 20, 23, 26, 29, 32, 35, 38
# affective schizotypy: 1, 28, 13, 19, 34, 
# social schizotypy: 4R, 7, 10R, 16, 22, 25R, 31, 37R

# negative schizotypy: 1, 4R, 7, 10R, 13, 16, 19, 22, 25R, 28, 31, 34, 37R


# disorganized schizotypy: 3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 33, 36

dat_cfa_mss <- 
  dat_cleaned %>% 
  dplyr::select(id_esm, contains("MSS_BH")) %>%
  distinct()
  
cfa_mss <- '


mss_pos_score =~  MSS_BH_8 +  MSS_BH_11 +  MSS_BH_14  +
                  MSS_BH_20 + MSS_BH_26 +  MSS_BH_29 +
                  MSS_BH_23 + MSS_BH_32 + MSS_BH_17 + MSS_BH_35 + MSS_BH_38 + MSS_BH_2 # +MSS_BH_5
            
mss_neg_score =~  MSS_BH_19 + MSS_BH_28 + MSS_BH_13 + 
                  MSS_BH_34 + MSS_BH_1 + MSS_BH_4 + MSS_BH_7 + # MSS_BH_10 + 
                  MSS_BH_16 + MSS_BH_25 + MSS_BH_31 + MSS_BH_37 + MSS_BH_22

mss_dis_score =~  MSS_BH_3 + MSS_BH_6 + MSS_BH_9 +
                  MSS_BH_12 + MSS_BH_15 + MSS_BH_18 +
                  MSS_BH_21 + MSS_BH_24 + MSS_BH_27 +
                  MSS_BH_30 + MSS_BH_36 + MSS_BH_33 
                  
'

fit_mss <- cfa(model = cfa_mss, # ordered = T,
                                ordered = c('MSS_BH_1', 'MSS_BH_2',
                                            'MSS_BH_3', 'MSS_BH_4',
                                            'MSS_BH_5', 'MSS_BH_6',
                                            'MSS_BH_7', 'MSS_BH_8',
                                            'MSS_BH_9', 'MSS_BH_10',
                                            'MSS_BH_11', 'MSS_BH_12',
                                            'MSS_BH_13', 'MSS_BH_14',
                                            'MSS_BH_15', 'MSS_BH_16',
                                            'MSS_BH_17', 'MSS_BH_18',
                                            'MSS_BH_19', 'MSS_BH_20',
                                            'MSS_BH_21', 'MSS_BH_22',
                                            'MSS_BH_23', 'MSS_BH_24',
                                            'MSS_BH_25', 'MSS_BH_26',
                                            'MSS_BH_27', 'MSS_BH_28',
                                            'MSS_BH_29', 'MSS_BH_30',
                                            'MSS_BH_31', 'MSS_BH_32',
                                            'MSS_BH_33', 'MSS_BH_34',
                                            'MSS_BH_35', 'MSS_BH_36',
                                            'MSS_BH_37', 'MSS_BH_38'),
                data = dat_cfa_mss, std.lv = TRUE, estimator = "WLSMV")

```

```{r}
fitMeasures(fit_mss, c("chisq", "df", "pvalue", "cfi", "tli",
                   "rmsea", "srmr"))
```

```{r}
modificationindices(fit_mss) %>%
  select("lhs", "op", "rhs", "mi") %>%
  arrange(desc(mi)) %>%
 # head(n = 10) %>%
  print()
```

```{r}

summary(fit_mss, fit.measures = TRUE, standardize = TRUE)

```


```{r}
# reliability of factors (omega)
(semTools::reliability(fit_mss))

# factor loadings
inspect(fit_mss, what = "std")$lambda

```

```{r, warning = T}

dat_cfa_mss_scores <-
  dat_cfa_mss %>%
  select(id_esm) %>%
  distinct() %>%
  cbind(lavPredict(fit_mss, newdata = dat_cfa_mss[2:39]))

dat_cleaned <-
  dat_cleaned %>% 
  left_join(dat_cfa_mss_scores, by = c("id_esm"))


```

```{r}
mzq_model <- 
  
  
'     mzq_refuse_reflection_score =~  mzq_1 + mzq_2 + mzq_4 # + mzq_3 
            
      mzq_emotional_awareness_score =~  mzq_5 + mzq_6 + mzq_7 + mzq_8

      mzq_psychic_equivalence_score =~  mzq_9 + mzq_10 + mzq_11 + mzq_12

      mzq_regulate_affect_score =~  mzq_13 + mzq_14 + mzq_15

      mzq_sum_score =~ mzq_refuse_reflection_score + mzq_emotional_awareness_score + 
      mzq_psychic_equivalence_score + mzq_regulate_affect_score

'

fit_mzq <- cfa(model = mzq_model, data = dat_mzq, std.lv = TRUE)
                              
                
```


```{r}

# dat_cleaned <-
#   dat_cleaned %>%
#   arrange(id_esm, day, beep) %>%
#   group_by(id_esm, day) %>%
#   mutate_at(vars(negaff_cfa,
#                  ambivalence_centered, 
#                  negaff_cfa_centered,
#                  posaff_cfa_centered,
#                  ple_cfa_centered,
#                  event_unpleasantness,
#                  event_pleasantness_centered,
#                  event_unpleasantness_centered,
#                  gonogo_dprime_centered,
#                  nback_dprime_centered,
#                  gonogo_corr_rt_median_centered,
#                  nback_corr_rt_median_centered),
#             .funs = list(lag = ~ lag(.))) %>%
#   ungroup() %>% 
#   distinct()

# dat_cleaned <-
#   dat_cleaned %>%
#   group_by(id_esm) %>%
#   mutate_at(vars(ambivalence,
#                  negaff_cfa,
#                  posaff_cfa,
#                  ple_cfa,
#                  event_pleasantness,
#                  event_unpleasantness,
#                  gonogo_dprime,
#                  nback_dprime,
#                  gonogo_corr_rt_median,
#                  nback_corr_rt_median),
#             .funs = list(mean = ~ mean(., na.rm = T))) %>%
#   ungroup()

```

centering variables, calculating 3days mean of momentary perceived distress, recoding objectified distressors

```{r}


dat_cleaned <-
  dat_cleaned %>%
  mutate(
    diary_counter =
      case_when(
        day <= 3  ~ 1,
        day <= 6  ~ 2,
        day <= 9  ~ 3,
        day <= 12 ~ 4,
        day <= 15 ~ 5,
        day <= 18 ~ 6,
        day <= 21 ~ 7,
        day <= 24 ~ 8,
        day <= 27 ~ 9,
        day <= 30 ~ 10,
      )
    )
# 
# dat_cleaned <- 
#   dat_cleaned %>%
#   group_by(id_esm, diary_counter) %>%
#   fill(c(dep_sum, dep_within, diary_counter, date_diary), .direction = "updown") %>% 
#   ungroup() %>% 
#   group_by(id_esm) %>%
#   fill(dep_between, .direction = "updown") %>% 
#   ungroup()
 
dat_cleaned <- 
  dat_cleaned %>% 
  mutate(stressor_mean = rowMeans(select(., c(stressor_1:stressor_19, stressor_21)), na.rm = T)) %>% 
  group_by(id_esm) %>%
  mutate_at(vars(stressor_sum, stressor_mean, ple_cfa, event_unpleasantness),
            .funs = list(centered = ~ scale(., scale = F, center = T))) %>% 
  distinct() %>% 
  ungroup() %>%
  group_by(id_esm, diary_counter) %>%
  mutate(event_unpleasantness_3days_mean = mean(event_unpleasantness, na.rm = T)) %>% 
  ungroup() %>% 
  mutate(stressor_sum_centered_int = stressor_sum_centered * -1,
         stressor_mean_centered_int = stressor_mean_centered * -1)
```

# 5: creation of final dataset for modeling

```{r}

# data for beep models - select model-relevant variables  

model_dat <-
  dat_cleaned %>%
  select(id_esm, session_esm, created_esm, edu, gender, age, city, start_date, date_diary, ended_diary, day, beep, diary_counter, three_day_block,

         event_unpleasantness, event_unpleasantness_centered, 
         
         ple_cfa, ple_cfa_centered, ple_between, 
         ple_control, ple_strange, ple_thoughts, ple_suspicion,
         ple_mistreated = ple_treatment, ple_perception, ple_thougtscontrol, ple_familiarstrange, ple_sum,

         contains(c("_mean", "stressor")),
         mss_dis, mss_pos, mss_neg, 
         mss_dis_score, mss_pos_score, 
         mss_neg_score, 
         ) %>%
  distinct()

```

```{r}

write.csv(model_dat, "data_processed/model_dat.csv", row.names = F)

```
