==> 01_stressor_cfa.Rmd <==
---
title: "01_stressor_cfa"
author: "Levente Ronai"
date: "2024-04-08"
output: 
  html_document:
   theme: spacelab
   code_download: true
   toc: true
   toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Setup

## load packages

```{r}
library(tidyverse)
library(lavaan)
library(semTools)

options(scipen = 999)
```

## import data

```{r}
big_data <- read.csv("data_processed/big_data.csv")

dat <- 
  big_data %>% 
  select(id_esm, session_esm, created_esm, age, gender, edu, city, start_date, date_esm, date_diary, three_day_block,
         day, beep, event_pleasantness, sad, afraid, unrested, angry, irritated, worried,
         contains(c("mss_", "ple_", "stressor", "SCL_90_depr_diary_")), -contains(c("minutes")))
```

## prep data

```{r}
dat_diary <-
  dat %>% 
  group_by(id_esm, three_day_block) %>% 
  mutate(
    across(
      c(angry, sad, irritated, worried, afraid,
        ple_strange, ple_thoughts, ple_suspicion, ple_treatment, event_pleasantness,
        ple_perception, ple_thougtscontrol, ple_familiarstrange),
      ~mean(., na.rm = T),
      .names = "{.col}_diarymean"
    )
  ) %>% 
  ungroup() %>% 
  distinct(id_esm, three_day_block, .keep_all = T)

```

# exploration

```{r}
dat_diary %>% 
  select(stressor_1:stressor_19, stressor_21) %>% 
  pivot_longer(stressor_1:stressor_21, names_to = "stressor_no", values_to = "score") %>% 
  ggplot(aes(score)) + 
  geom_histogram() + 
  facet_wrap(~stressor_no)

```


```{r}
dat_diary %>% 
    pivot_longer(stressor_1:stressor_18, names_to = "stressor_no", values_to = "score") %>% 
  count(stressor_no, score) %>% 
  filter(!is.na(score)) %>% 
  arrange(stressor_no)
```

## reflective models

### pca with the stressor variables that show some variance

```{r}
prin <-
  dat %>%
  select(stressor_9, stressor_10, stressor_11, stressor_12, stressor_13, stressor_15, stressor_16, stressor_17) %>% 
    filter_at(vars(stressor_9:stressor_17), all_vars(!is.na(.)))


(parallel <- psych::fa.parallel(prin, fm = "pa"))

(principal <-
    psych::principal(prin, nfactors = 3, rotate = "oblimin"))
```

### confirmatory factor analysis with the stressor variables that show some variance

#### correlated factors

```{r}
#### hierarchical factors


dat_cfa_stressor <-
  dat_diary %>% 
  select(id_esm, three_day_block, stressor_9, stressor_10, stressor_11,
         stressor_12, stressor_13, stressor_15, stressor_16, stressor_17) %>%
  filter_at(vars(stressor_9:stressor_17), all_vars(!is.na(.))) 
  

cfa_stressor_hierarchical_syntax <- '

level: 1
      stressor_I  =~ stressor_9 +  stressor_10 +  stressor_11
      
      stressor_II  =~ stressor_12 +  stressor_13 # + stressor_17
      
      stressor_III  =~ stressor_15 +  stressor_16
      
      stressor_higher  =~ stressor_I + stressor_II + stressor_III
    

      
level: 2
      stressor_I_b  =~ stressor_9 +  stressor_10 +  stressor_11
      
      stressor_II_b  =~ stressor_12 +  stressor_13 + stressor_11
      
      stressor_III_b  =~ stressor_15 +  stressor_16
      
    #  stressor_higher_b  =~ stressor_I_b + stressor_II_b + stressor_III_b

      
'

fit_stressor_hierarchical <- cfa(model = cfa_stressor_hierarchical_syntax, data = dat_cfa_stressor, std.lv = TRUE, cluster = "id_esm", estimator = "MLR")
```

```{r, warning = T}

fitMeasures(fit_stressor_hierarchical, c("chisq", "df", "pvalue", "cfi", "tli", "rmsea", "srmr_within", "srmr_between"))
```

```{r}
summary(fit_stressor_hierarchical, std = T)

```

```{r}
modificationindices(fit_stressor_hierarchical) %>% arrange(-mi)

```


```{r}
# within/betwwe-person factor loadings
(inspect(fit_stressor_hierarchical, what = "std")$within$lambda)
(inspect(fit_stressor_hierarchical, what = "std")$id_esm$lambda)
```
```{r}

compRelSEM(fit_stressor_hierarchical, tau.eq = F, dropSingle = F)

```


```{r}
# reliability
(semTools::reliability(fit_stressor_hierarchical))

```
