==> 02_modeling.Rmd <==
---
title: "levente_data_process"
author: "Levente Ronai"
output: html_document
  html_document:
   theme: spacelab
   code_download: true
   toc: true
   toc_float: true
editor_options: 
  chunk_output_type: inline
---

```{r, message=F, warning=F, include=F}

options(scipen = 999)

library(corrplot)
library(readxl)
library(xts)
library(psych)
library(qgraph)
library(tidyverse)
library(dplyr)
library(purrr)
library(lme4)
library(lmerTest)
library(lavaan)
library(sjPlot)
library(ggplot2)
library(car)
library(mlVAR)
library(graphicalVAR)
library(broom)
library(performance)
# library(confintr)
# library(goft)

```


# import data

```{r, message=F, warning=F}

dat_cleaned <- read_csv("data_processed/dat_cleaned.csv")

model_dat <- read.csv("data_processed/model_dat.csv")

# model_dat <- 
#   model_dat %>%
#   filter_at(vars(event_unpleasantness_centered:ple_cfa_centered, negaff_cfa_centered_lag), all_vars(!is.na(.)))

```


It has been suggested that some non-schizophrenic individuals may exhibit PLEs frequently, but in a fashion that is not distressing but life-enhancing (“benign schizotypy”). In schizophrenia and at-risk individuals, however, psychotic symptoms are not only distressing, but also triggered stress.

Results suggest two discrete effects: On the one hand, individuals high in disorganised and negative schizotypy showed stress-dependent increases in PLEs; without added effects of positive schizotypy. On the other, individuals low in negative and disorganised schizotypy showed higher levels of PLEs solely as a function of positive schizotypy but not stress.

We discuss these findings in light of the fully-dimensional model of schizotypy and hypothesize that PLEs in individuals high in schizotypy-facets suggested to convey risk-for-schizophrenia (negative and disorganised) may reflect qualitatively different entities than PLEs in individuals with low values in these facets, but high expressions of positive schizotypy (“happy schizotypes”). Additionally, we emphasize the importance of not overlooking the disorganised schizotypy-facet in related research.

## Working memory and response inhibition

```{r}
model_dat %>% 
  pivot_longer(names_to = "cog_task", cols = c(nback_dprime, gonogo_dprime), values_to = "dprime" ) %>% 
  group_by(day, id_esm, cog_task) %>% 
  summarise(dprime = mean(dprime, na.rm = T)) %>%   
  ggplot(aes(day, dprime)) + 
  geom_jitter(aes(color = cog_task), alpha = .3) + 
  geom_line(aes(group = id_esm), alpha = .2) + 
  geom_smooth() + 
  lims(x = c(1, 20)) + 
  facet_wrap(~cog_task)
```
```{r}
model_dat %>% 
  pivot_longer(names_to = "cog_task", cols = c(nback_dprime, gonogo_dprime), values_to = "dprime" ) %>% 
  group_by(beep, id_esm, cog_task) %>% 
  summarise(dprime = median(dprime, na.rm = TRUE)) %>% 
  ggplot(aes(beep, dprime)) + 
  geom_line(aes(group = id_esm), alpha = .2) + 
  geom_jitter(aes(color = cog_task), alpha = .2) + 
  stat_summary(fun = "median", geom = "line") +
  facet_wrap(~cog_task)
```

### Links with schizotypy

```{r}
# correlation between trait self report and mean state behavioural measures of cognitive performance

corr_cog <- 
  model_dat %>% 
  dplyr::select(id_esm, nback_dprime_grand_centered, gonogo_dprime_grand_centered, mss_dis_score) %>% 
  distinct()

shapiro.test(corr_cog$nback_dprime_grand_centered)
shapiro.test(corr_cog$gonogo_dprime_grand_centered)
shapiro.test(corr_cog$mss_dis_score)
# normality not assumed

cor.test(corr_cog$mss_dis_score, corr_cog$nback_dprime_grand_centered, method = "spearman")
cor.test(corr_cog$mss_dis_score, corr_cog$gonogo_dprime_grand_centered, method = "spearman")

# ci_cor(corr_cog, probs = c(0.025, 0.975), method = "spearman", type = "bootstrap", boot_type = "basic")
# 
# corr_cog <-
#   corr_cog %>%
  # mutate(mss_dis_score_log = log10(mss_dis_score + 2),
  #        ple_between_log = log10(ple_between + 2))

# ggplot(corr_cog, aes(mss_dis_score_sqrt, nback_dprime_grand_centered)) +
#   geom_point(alpha = 0.5, position = "jitter") +
#   geom_smooth(method = "lm", fill = "#8fb9dc", color = "#609CCE") +
#   theme_classic() +
#   # labs(x = "Trait rumination",
#   #      y = "Mean state rumination"
#   # ) +
#   theme(axis.title = element_text(size = 12))

# ggplot(corr_cog, aes(mss_dis_score, nback_dprime_grand_centered)) +
#   geom_point(alpha = 0.5, position = "jitter") +
#   geom_smooth(method = "lm", fill = "#8fb9dc", color = "#609CCE") +
#   theme_classic() +
# 
#   theme(axis.title = element_text(size = 12))
```

```{r}
mod_wm_scht <- 
  model_dat %>% 
  lmer(
    nback_dprime ~
      poly(day,3 ) + 
      # poly(beep, 2) +
      age + 
      gender *
      (mss_dis_score + mss_pos_score + mss_neg_score) +
      (1 | id_esm), 
    data = . 
  )

summary(mod_wm_scht)
# plot_model(mod_wm_scht, type = "int")
```


```{r}
mod_inh_scht <- 
  model_dat %>% 
  lmer(
    gonogo_dprime ~
      poly(day, 3) + poly(beep, 2) +
      # day and beep seem to have nonlinear effects
      age + 
      gender *
      (mss_dis_score + mss_pos_score + mss_neg_score) + 
      (1 | id_esm), 
    data = . 
  )

summary(mod_inh_scht)
```
### Links with PLE

```{r}
mod_ple_wm_inh <- 
  model_dat %>% 
    lmer(
      ple_cfa ~  
        poly(day, 3) + poly(beep, 2) +
        age + 
        gonogo_dprime_centered + gonogo_dprime_grand_centered + 
        nback_dprime_centered + nback_dprime_grand_centered +
    ( gonogo_dprime_centered | id_esm),
    data = .
    )

summary(mod_ple_wm_inh)

```



```{r}

# model_dat <- 
#   model_dat %>% 
#   mutate(ple_log = log10(ple_cfa+3),
#          negaff_log = log10(negaff_cfa+3))

```

## Psychotic reactivity

```{r}
m1 <- 
  lmer(
    ple_cfa ~ 
      day + beep + gender + age +
      event_unpleasantness_centered * mss_neg_score  +
      event_unpleasantness_centered * mss_pos_score  +
     # event_unpleasantness_centered * mss_dis_score  +

    #  ple_cfa_centered_lag + 

      (event_unpleasantness_centered | id_esm), data = model_dat)


tab_model(m1, show.aicc = T, show.std = T)
plot_model(m1, type = "int")
```


```{r}
m2 <- 
  lmer(
    ple_cfa ~ 
      day + beep + gender + age +
      event_unpleasantness_centered * mss_neg_score  +
      event_unpleasantness_centered * mss_pos_score  +
      event_unpleasantness_centered * mss_dis_score  +

#      ple_cfa_centered_lag + 

      (event_unpleasantness_centered | id_esm), data = model_dat)


tab_model(m2, show.aicc = T, show.std = T)
plot_model(m2, type = "int")
```


```{r}
m3 <- 
  lmer(
    ple_cfa ~ 
      day + beep + gender + age +
      event_unpleasantness_centered * mss_neg_score  +
      event_unpleasantness_centered * mss_pos_score  +
      event_unpleasantness_centered * mss_dis_score  +
      ple_cfa_centered_lag + 

      (event_unpleasantness_centered + ple_cfa_centered_lag | id_esm), data = model_dat)


tab_model(m3, show.aicc = T, show.std = T)
plot_model(m3, type = "int")

```

```{r}
plot_model(m1, show.values = TRUE, value.offset = .3, type = "std",
           show.p = TRUE, value.size = 3, line.size = 0.5, dot.size = 1, digits = 3) + ylim(-.4, .4)

```

```{r}
plot_model(m2, show.values = TRUE, value.offset = .3, type = "std",
           show.p = TRUE, value.size = 3, line.size = 0.5, dot.size = 1, digits = 3) + ylim(-.4, .4)

```

```{r}
plot_model(m3, show.values = TRUE, value.offset = .3, type = "std",
           show.p = TRUE, value.size = 3, line.size = 0.5, dot.size = 1, digits = 3) + ylim(-.4, .4)

```

```{r}
# m_na <- 
#   lmer(
#     negaff_cfa ~ 
#       day + beep + gender + age +
#       event_unpleasantness_centered * mss_dis_score  +
#       event_unpleasantness_centered * mss_pos_score  +
#       event_unpleasantness_centered * mss_neg_score  +
#       (event_unpleasantness_centered| id_esm), data = model_dat)
# 
# 
# tab_model(m_na)
# plot_model(m_na, type = "int")
```


```{r}
model_dat %>% 
  select(id_esm, gender, age, day, beep, event_unpleasantness_centered, ple_cfa, mss_dis, mss_pos, mss_neg) %>% 
  filter(!is.na(ple_cfa)) %>%
  select(id_esm, gender, age) %>% 
  distinct() %>%
  summarise(table(gender), median(age))
```



```{r}
m <- 
  lmer(
    ple_cfa ~ 
      day + beep + gender + age +
      # nback_dprime_centered * event_unpleasantness_centered +
      # nback_dprime_grand_centered  * event_unpleasantness_centered +
      mss_dis_score * event_unpleasantness_centered +
      (event_unpleasantness_centered | id_esm), data = model_dat)


m0 <- 
  lmer(
    ple_cfa ~ 
      day + beep + gender + age +
      gonogo_dprime_centered * event_unpleasantness_centered +
      gonogo_dprime_grand_centered * event_unpleasantness_centered +
      mss_dis_score * event_unpleasantness_centered +
      (gonogo_dprime_centered + event_unpleasantness_centered | id_esm), data = model_dat)


tab_model(m0)

plot_model(m, type = "diag")

```



```{r}
plot_model(m0, show.values = TRUE, value.offset = .3, # type = "std",
           show.p = TRUE, value.size = 3, line.size = 0.5, dot.size = 1, digits = 3) + ylim(-.1, .1)

```


```{r}
m1 <- 
  lmer(
    ple_cfa ~
      day + beep + gender + age +
      nback_dprime_centered * negaff_cfa_centered +
      nback_dprime_grand_centered * negaff_cfa_centered +
      mss_dis_score  * negaff_cfa_centered +
      (nback_dprime_centered * negaff_cfa_centered  | id_esm), data = model_dat)


m2 <- 
  lmer(
    ple_cfa ~ 
      day + beep + gender + age +
      gonogo_dprime_centered * negaff_cfa_centered +
      gonogo_dprime_grand_centered  * negaff_cfa_centered +
      mss_dis_score  * negaff_cfa_centered +
      (gonogo_dprime_centered * negaff_cfa_centered | id_esm), data = model_dat)


tab_model(m1, m2)

plot_model(m1, type = "int")
plot_model(m2, type = "int")
```


```{r, fig.width=7, fig.height=7}
check_model(m1)
```

```{r}
plot_model(m1, show.values = TRUE, value.offset = .3, # type = "std",
           show.p = TRUE, value.size = 3, line.size = 0.5, dot.size = 1, digits = 3) + ylim(-.1, .1)

```

```{r}
plot_model(m2, show.values = TRUE, value.offset = .3, # type = "std",
           show.p = TRUE, value.size = 3, line.size = 0.5, dot.size = 1, digits = 3) + ylim(-.1, .1)

```

## Affect dynamics and schizotypy 

### Positive and negative schizotypy

```{r}
affdyn_pos_neg <- 
  model_dat %>%   
  lmer(
    negaff_cfa ~ 
      day + beep + gender + age + (negaff_cfa_centered_lag) * (mss_pos_score + mss_neg_score) +  
      (ple_cfa_centered | id_esm), 
    data = .)


tab_model(affdyn_pos_neg, show.aicc = T, show.std = T)
plot_model(affdyn_pos_neg, type = "int")

```

### Disorg schizotypy

```{r}
affdyn_dis <- 
  model_dat %>%   
  lmer(
    negaff_cfa ~ 
      day + beep + gender + age + (negaff_cfa_centered_lag) * (mss_dis_score) +  
      (ple_cfa_centered | id_esm), 
    data = .)


tab_model(affdyn_dis, show.aicc = T, show.std = T)
plot_model(affdyn_dis, type = "int")

```

### Positive and negative and disorg schizotypy

```{r}
affdyn_pos_neg_dis <- 
  model_dat %>%   
  lmer(
    negaff_cfa ~ 
      day + beep + gender + age + (negaff_cfa_centered_lag) * (mss_dis_score + mss_pos_score + mss_neg_score) +  
      (ple_cfa_centered | id_esm), 
    data = .)


tab_model(affdyn_pos_neg_dis, show.aicc = T, show.std = T)
plot_model(affdyn_pos_neg_dis, type = "int")

```

- positive and disorganised schizotypy predict higher negative affect
- negative scht predicts lower negative affect if disorg is controlled for
- positive and negative schizotypy both independently predict stronger negative affective inertia
- without adjusting for positive and negative schizotypy, disorg shows the same effect
- however, when all schizotypy dimensions are entered, disorg predicts reduced inertia
  - inspecting the plots suggests that pos and negative schizotypy held constant, high disorganisation predicts a high negative affect, even when preceding negative affect was low (in high disorganisation, mood tends to go to negative, no matter what happened before), which effectively reduces autocorrelation


## Ambivalence

```{r}
ambivalence_model <-

  lmer(
    ambivalence ~ 
      day + beep + gender + age +
      event_unpleasantness_centered * ple_cfa_centered + 
      negaff_cfa + I(negaff_cfa^2) +
      (event_unpleasantness_centered * ple_cfa_centered | id_esm), data = model_dat)

tab_model(ambivalence_model)
```


```{r}
plot_model(ambivalence_model, type = "diag")
```

```{r}
ambivalence_model_trait <-

  lmer(
    ambivalence ~ 
      day + beep + gender + age +
       event_unpleasantness_centered * mss_dis_score + negaff_cfa + I(negaff_cfa^2) +
      (event_unpleasantness_centered  | id_esm), data = model_dat)

summary(ambivalence_model_trait)
```


```{r}
plot_model(ambivalence_model_trait, type = "int")
```

```{r}

bias_model <-
  
  model_dat %>% 
#  filter(z_event_within_ind == "Neutral") %>% 
  
  lmer(
    negaff_cfa ~ 
      day + beep + gender + age +
      event_unpleasantness_centered * ple_cfa_centered + 
      (event_unpleasantness_centered * ple_cfa_centered  | id_esm), data = .)



summary(bias_model)


plot_model(bias_model, type = "int")

```


```{r}

offset_model <-
  
  model_dat %>% 
#  filter(z_event_within_ind == "Neutral") %>% 
  
  lmer(
    posaff_cfa ~ 
      day + beep + gender + age + 
      event_unpleasantness_centered * ple_cfa_centered +
      (event_unpleasantness_centered * ple_cfa_centered | id_esm), data = .)



summary(offset_model)


plot_model(offset_model, type = "pred")

```


```{r, warning=F}

mediation_model <- '
        
        level: 1
        
              negaff_within =~
              sad + afraid + unrested + irritated + angry
              
              ple_within =~ 
              ple_control + ple_strange + ple_thoughts + 
              ple_suspicion + ple_treatment + ple_thougtscontrol

            # direct effects
              ple_within  ~
              d_w * negaff_within +       
              # control variables
              beep + day
                
            # mediator 1
              gonogo_dprime ~  
              m1_w * negaff_within +
              # control variables
              beep + day
                
             # mediator 2
              nback_dprime ~  
              m2_w * negaff_within +
              # control variables
              beep + day
            
            # indirect effects
              na_gonog_ple_w :=  m1_w * d_w
              na_nback_ple_w :=  m2_w * d_w
             
            # total effects
              total_ple_w :=  
              d_w + (m1_w * d_w) + (m2_w * d_w)
            
        level: 2
            
            negaff_between =~
              sad + afraid + unrested + irritated + angry
            
            ple_between =~ 
              ple_control + ple_strange + ple_thoughts + 
              ple_suspicion + ple_treatment + ple_thougtscontrol
            
            # direct effects
              ple_between  ~
              d_b * negaff_between +       
              # control variables
              age + gender
            
            # mediator 1
              gonogo_dprime ~  
              m1_b * negaff_between +
              # control variables
              age + gender
              
            # mediator 2
              nback_dprime ~  
              m2_b * negaff_between +
              # control variables
              age + gender
            
            # indirect effects
              ple_gonog_na_b :=  m1_b * d_b
              ple_nback_na_b :=  m2_b * d_b
            
            # total effects
              total_ple_b :=  
              d_b + (m1_b * d_b) + (m2_b * d_b)
            '

fit <- sem(model = mediation_model, data = model_dat, cluster = "id_esm", estimator = "ML")

```

```{r}

summary(fit, std = T, fit.measures = T)

```


```{r, include=F}
# nback_dprime_centered

# mod_nback_pca <-
# 
#    lmer(
#      negaff ~ poly(beep, 2) + day + age + gender +
#       nback_dprime_centered * event_unpleasantness_centered +
#       negaff_centered_lag + 
#       (event_unpleasantness_centered + nback_dprime_centered + negaff_centered_lag | id_esm ), data = model_dat)


# vif(mod_nback_pca)
# summary(mod_nback_pca)

```


```{r, message=F, warning=F, include=F}

vars <- c( "sad", "afraid", "unrested", "irritated", "angry", "ple_control", "ple_strange", "ple_thoughts", "ple_suspicion", "ple_treatment", "ple_thougtscontrol", #"gonogo_dprime", "nback_dprime",
           "event_unpleasantness")


vars <- c("ple_control", "ple_strange", "ple_thoughts", "ple_suspicion", "ple_treatment", "ple_thougtscontrol", "gonogo_dprime", "nback_dprime", "event_unpleasantness")


```


```{r, include=F}  

#, echo=FALSE, message = FALSE, warning=F, include=F}

res1_corr <-
  mlVAR(
    data = model_dat,
    vars = vars,
    idvar = "id_esm",
    dayvar = "day",
    beepvar = "beep",
    contemporaneous = "correlated",
    temporal = "correlated")

```


```{r,message=F, warning=F, include=F}

plot(res1_corr, "temporal", title = "Temporal", edge.labels = T, label.scale.equal = T, label.cex = 2.8, nonsig = "hide" , cut = 0.2)

```

```{r, message=F, warning=F}

plot(res1_corr, "contemporaneous", title = "Contemporaneous", edge.labels = T, label.scale.equal = T, label.cex = 2.8, nonsig = "hide" , cut = 0.2)

```

```{r, message=F, warning=F}

plot(res1_corr, "between", title = "between", edge.labels = T, label.scale.equal = T, label.cex = 2.8, nonsig = "hide" , cut = 0.2)

```
