==> 02_stat_modeling_ple_react.Rmd <==
---
title: "02_stat_modeling_ple_react"
author: "Levente Ronai"
output: html_document
  html_document:
   theme: spacelab
   code_download: true
   toc: true
   toc_float: true
editor_options: 
  chunk_output_type: inline
---

```{r, include=F}

options(scipen = 999)

suppressPackageStartupMessages(library(readxl))
suppressPackageStartupMessages(library(xts))
suppressPackageStartupMessages(library(psych))
suppressPackageStartupMessages(library(qgraph))
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(lme4))
suppressPackageStartupMessages(library(lmerTest))
suppressPackageStartupMessages(library(sjPlot))
suppressPackageStartupMessages(library(car))
suppressPackageStartupMessages(library(broom))
suppressPackageStartupMessages(library(performance))
suppressPackageStartupMessages(library(lmeresampler))

```

import data

```{r}
dat_cleaned <- read_csv("data_processed/dat_cleaned.csv")

model_dat <- read.csv("data_processed/model_dat.csv")
```
testing the correlation between stressor-expsoure and stress-appraisal (event-unpleasantness)

```{r}
library(rstatix)
# correlation between stressor and event unpleasantness
model_dat %>% 
  group_by(id_esm) %>%
  summarise(
    event_unpleasantness = mean(event_unpleasantness, na.rm = T),
    stressor_mean = mean(stressor_mean, na.rm = T)*-1) %>%
  select(event_unpleasantness, stressor_mean) %>%
  cor_test(method = "spearman")
```

testing predictive validity of stressor scale

```{r}

# m_baseline_ple <-
#   lmer(
#     ple_cfa ~ age + gender + day + beep +
#       stressor_mean_centered_int + (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# m_baseline_event <- 
#   lmer(
#     event_unpleasantness ~ age + gender + day + beep + 
#       stressor_mean_centered_int + (stressor_mean_centered_int | id_esm), data = model_dat)
# saveRDS(m_baseline_ple, "models/m_baseline_ple.rds")
# saveRDS(m_baseline_event, "models/m_baseline_event.rds")

m_baseline_ple <- readRDS("models/m_baseline_ple.rds")
m_baseline_event <- readRDS("models/m_baseline_event.rds")

tab_model(m_baseline_event,
          show.aicc = T, show.std = T)

```

### ~ ple

```{r}

# m1 <- 
#   lmer(
#     ple_cfa ~ 
#       day + beep + gender + age +
#       stressor_mean_centered_int * mss_neg_score  +
#       #     stressor_mean_centered_int * mss_pos_score  +
#       #     stressor_mean_centered_int * mss_dis_score  +
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# m2 <-
#   lmer(
#     ple_cfa ~
#       day + beep + gender + age +
#       stressor_mean_centered_int * mss_neg_score  +
#       stressor_mean_centered_int * mss_pos_score  +
#       #     stressor_mean_centered_int * mss_dis_score  +
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# m3 <-
#   lmer(
#     ple_cfa ~
#       day + beep + gender + age +
#       stressor_mean_centered_int * mss_neg_score  +
#       stressor_mean_centered_int * mss_pos_score  +
#       stressor_mean_centered_int * mss_dis_score  +
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# 
# m4 <-
#   lmer(
#     ple_cfa ~
#       day + beep + gender + age +
#       # stressor_mean_centered_int * mss_neg_score  +
#       # stressor_mean_centered_int * mss_pos_score  +
#       stressor_mean_centered_int * mss_dis_score  +
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# m5 <-
#   lmer(
#     ple_cfa ~
#       day + beep + gender + age +
#       # stressor_mean_centered_int * mss_neg_score  +
#       stressor_mean_centered_int * mss_pos_score  +
#       # stressor_mean_centered_int * mss_dis_score  +
#       (stressor_mean_centered_int | id_esm), data = model_dat)

# saveRDS(m1, "models/m1.rds")
# saveRDS(m2, "models/m2.rds")
# saveRDS(m3, "models/m3.rds")
# saveRDS(m4, "models/m4.rds")
# saveRDS(m5, "models/m5.rds")

m1 <- readRDS("models/m1.rds")
m2 <- readRDS("models/m2.rds")
m3 <- readRDS("models/m3.rds")

tab_model(m1,
          m2, 
          m3,
          show.aicc = T, show.std = T)

```

```{r}
plot_model(m3, type = "diag")
vif(m3)
```

### ~ event unpleasantness

```{r}

# m7 <- 
#   lmer(
#     event_unpleasantness ~ 
#       day + beep + gender + age +
#       stressor_mean_centered_int * mss_neg_score +
#       # stressor_mean_centered_int * mss_pos_score +
#       #stressor_mean_centered_int * mss_dis_score + 
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# 
# m8 <- 
#   lmer(
#     event_unpleasantness ~ 
#       day + beep + gender + age +
#       stressor_mean_centered_int * mss_neg_score +
#       stressor_mean_centered_int * mss_pos_score +
#       #      stressor_mean_centered_int * mss_dis_score + 
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# m9 <- 
#   lmer(
#     event_unpleasantness ~ 
#       day + beep + gender + age +
#       stressor_mean_centered_int * mss_neg_score +
#       stressor_mean_centered_int * mss_pos_score +
#       stressor_mean_centered_int * mss_dis_score + 
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# 
# m10 <-
#   lmer(
#     event_unpleasantness ~
#       day + beep + gender + age +
#       # stressor_mean_centered_int * mss_neg_score +
#       # stressor_mean_centered_int * mss_pos_score +
#       stressor_mean_centered_int * mss_dis_score +
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# 
# m11 <-
#   lmer(
#     event_unpleasantness ~
#       day + beep + gender + age +
#       # stressor_mean_centered_int * mss_neg_score +
#       stressor_mean_centered_int * mss_pos_score +
#       #   stressor_mean_centered_int * mss_dis_score +
#       (stressor_mean_centered_int | id_esm), data = model_dat)
# saveRDS(m7, "models/m7.rds")
# saveRDS(m8, "models/m8.rds")
# saveRDS(m9, "models/m9.rds")
# saveRDS(m10, "models/m10.rds")
# saveRDS(m11, "models/m11.rds")

m7 <- readRDS("models/m7.rds")
m8 <- readRDS("models/m8.rds")
m9 <- readRDS("models/m9.rds")

tab_model(m7, m8, m9,
  show.aicc = T, show.std = T)


```


```{r}
plot_model(m9, type = "diag")
vif(m9)
```
