==> 03_new_bootstrapping_visu.Rmd <==
---
title: "03_new_bootstrapping_visu"
author: "Levente Ronai"
date: "2024-04-08"
output: 
  html_document:
   theme: spacelab
   code_download: true
   toc: true
   toc_float: true
editor_options: 
  chunk_output_type: inline
---

```{r, include=F}

options(scipen = 999)

suppressPackageStartupMessages(library(readxl))
suppressPackageStartupMessages(library(xts))
suppressPackageStartupMessages(library(psych))
suppressPackageStartupMessages(library(qgraph))
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(purrr))
suppressPackageStartupMessages(library(lme4))
suppressPackageStartupMessages(library(lmerTest))
suppressPackageStartupMessages(library(sjPlot))
suppressPackageStartupMessages(library(ggplot2))
suppressPackageStartupMessages(library(car))
suppressPackageStartupMessages(library(broom))
suppressPackageStartupMessages(library(performance))
suppressPackageStartupMessages(library(lmeresampler))
suppressPackageStartupMessages(library(ggpattern))
suppressPackageStartupMessages(library(patchwork))


```

import data

```{r}
dat_cleaned <- read_csv("data_processed/dat_cleaned.csv")


model_dat <- read.csv("data_processed/model_dat.csv")
```

```{r}

# m_baseline_ple <- readRDS("models/m_baseline_ple.rds")

# boot_m_baseline_ple <-
#   bootstrap(
#     m_baseline_ple,
#     .f = fixef,
#     type = "wild",
#     hccme =  "hc2",
#     aux.dist = "mammen",
#     B = 5000,
#     resample = c(TRUE, TRUE)
#   )
# 
# saveRDS(boot_m_baseline_ple, "models/boot_m_baseline_ple.rds")

boot_m_baseline_ple <- readRDS("models/boot_m_baseline_ple.rds")

```

```{r}
list_case_ple_baseline <- 
  list(boot_m_baseline_ple)
ci_boot_baseline_ple <-
  list_case_ple_baseline %>%
  map(confint, type = "norm")


as.data.frame(ci_boot_baseline_ple) %>% print.data.frame(., digits = 2)

```

```{r}

# m_baseline_event <- readRDS("models/m_baseline_event.rds")
# 
# boot_baseline_event <-
#   bootstrap(
#     m_baseline_event,
#     .f = fixef,
#     type = "wild",
#     hccme =  "hc2",
#     aux.dist = "mammen",
#     B = 5000 ,
#     resample = c(TRUE, TRUE)
#   )
# 
# saveRDS(boot_baseline_event, "models/boot_baseline_event.rds")

boot_baseline_event <- readRDS("models/boot_baseline_event.rds")

```

```{r}
list_case_event_baseline <- 
  list(boot_baseline_event)
ci_boot_baseline_event <-
  list_case_event_baseline %>%
  map(confint, type = "norm")
as.data.frame(ci_boot_baseline_event) %>% print.data.frame(., digits = 2)

```
```{r}

m2 <- readRDS("models/m2.rds")
# 
# boot_m2 <-
#   bootstrap(
#     m2,
#     .f = fixef,
#     type = "wild",
#     hccme =  "hc2",
#     aux.dist = "mammen",
#     B = 5000 ,
#     resample = c(TRUE, TRUE)
#   )
# 
# saveRDS(boot_m2, "models/boot_m2.rds")
# 
m8 <- readRDS("models/m8.rds")
# 
# boot_m8 <-
#   bootstrap(
#     m8,
#     .f = fixef,
#     type = "wild",
#     hccme =  "hc2",
#     aux.dist = "mammen",
#     B = 5000 ,
#     resample = c(TRUE, TRUE)
#   )
# saveRDS(boot_m8, "models/boot_m8.rds")

boot_m2 <- readRDS("models/boot_m2.rds")

boot_m8 <- readRDS("models/boot_m8.rds")
```

```{r}
list_case_m2 <- 
  list(boot_m2)
ci_boot_m2 <-
  list_case_m2 %>%
  map(confint, type = "norm")

(ci_boot_m2 <- 
  as.data.frame(ci_boot_m2)  %>% 
  mutate_at(vars(estimate, lower, upper), ~ round(., 3)) %>%
  print.data.frame(.))

# # knitr::kable(ci_boot_m2, format = "html")


```

```{r}
list_case_m8 <- 
  list(boot_m8)
ci_boot_m8 <-
  list_case_m8 %>%
  map(confint, type = "norm")

ci_boot_m8 <- 
  as.data.frame(ci_boot_m8) %>% 
  mutate_at(vars(estimate, lower, upper), ~ round(., 3)) %>%
  print.data.frame(.)


# knitr::kable(ci_boot_m8, format = "html")
```

```{r}

# m3 <- readRDS("models/m3.rds")
# 
# boot_m3 <-
#   bootstrap(
#     m3,
#     .f = fixef,
#     type = "wild",
#     hccme =  "hc2",
#     aux.dist = "mammen",
#     B = 5000 ,
#     resample = c(TRUE, TRUE)
#   )
# 
# saveRDS(boot_m3, "models/boot_m3.rds")

boot_m3 <- readRDS("models/boot_m3.rds")
```

```{r}
list_case_m3 <- 
  list(boot_m3)
ci_boot_m3 <-
  list_case_m3 %>%
    map(confint, type = "norm")

(ci_boot_m3 <- 
  as.data.frame(ci_boot_m3) %>% 
  mutate_at(vars(estimate, lower, upper), ~ round(., 3)) %>%
  print.data.frame(.))

# knitr::kable(ci_boot_m3, format = "html")
```

```{r}

# m9 <- readRDS("models/m9.rds")
# 
# boot_m9 <-
#   bootstrap(
#     m9,
#     .f = fixef,
#     type = "wild",
#     hccme =  "hc2",
#     aux.dist = "mammen",
#     B = 5000,
#     resample = c(TRUE, TRUE)
#   )
# 
# saveRDS(boot_m9, "models/boot_m9.rds")

boot_m9 <- readRDS("models/boot_m9.rds")
```

```{r}
list_case_m9 <- 
  list(boot_m9)

ci_boot_m9 <-
  list_case_m9 %>%
 map(confint, type = "norm")

(ci_boot_m9 <- 
  as.data.frame(ci_boot_m9)%>% 
  mutate_at(vars(estimate, lower, upper), ~ round(., 3)) %>%
  print.data.frame(.))


# knitr::kable(ci_boot_m9, format = "html")

```

```{r}
# extracting results into a tibble

boot_dat_mod_3 <- 
  list_case_m3 %>% 
  map("replicates") %>% 
  
map(
    ~tibble(
      Stress_Main = na.omit(.x$stressor_mean_centered_int),
      Disorganized_Main = na.omit(.x$mss_dis_score),
      Positive_Main = na.omit(.x$mss_pos_score),
      Negative_Main = na.omit(.x$mss_neg_score),
      Disorganized_Stress_Interaction = na.omit(.x$"stressor_mean_centered_int:mss_dis_score"),
      Positive_Stress_Interaction = na.omit(.x$"stressor_mean_centered_int:mss_pos_score"),
      Negative_Stress_Interaction = na.omit(.x$"stressor_mean_centered_int:mss_neg_score")
      
    )
  ) %>%
  bind_rows() %>% 
  mutate(main_diff_dispos = Disorganized_Main - Positive_Main,
         int_diff_dispos = Disorganized_Stress_Interaction - Positive_Stress_Interaction,
         main_diff_disneg = Disorganized_Main - Negative_Main,
         int_diff_disneg = Disorganized_Stress_Interaction - Negative_Stress_Interaction)

boot_dat_mod_9 <- 
  list_case_m9 %>% 
  map("replicates") %>% 
  
map(
    ~tibble(
      Stress_Main = na.omit(.x$stressor_mean_centered_int),
      Disorganized_Main = na.omit(.x$mss_dis_score),
      Positive_Main = na.omit(.x$mss_pos_score),
      Negative_Main = na.omit(.x$mss_neg_score),
      Disorganized_Stress_Interaction = na.omit(.x$"stressor_mean_centered_int:mss_dis_score"),
      Positive_Stress_Interaction = na.omit(.x$"stressor_mean_centered_int:mss_pos_score"),
      Negative_Stress_Interaction = na.omit(.x$"stressor_mean_centered_int:mss_neg_score")
      
    )
  ) %>%
  bind_rows() %>% 
  mutate(main_diff_dispos = Disorganized_Main - Positive_Main,
         int_diff_dispos  = Disorganized_Stress_Interaction - Positive_Stress_Interaction,
         main_diff_disneg = Disorganized_Main - Negative_Main,
         int_diff_disneg  = Disorganized_Stress_Interaction - Negative_Stress_Interaction)
  
```


```{r}
boot_dat_mod_3 <- 
  boot_dat_mod_3%>%
  select(-Stress_Main)
  
boot_dat_mod_long_3 <- 
  boot_dat_mod_3 %>%
    pivot_longer(c(
    Disorganized_Main, 
    Positive_Main,
    Negative_Main,
    Disorganized_Stress_Interaction,
    Positive_Stress_Interaction,
    Negative_Stress_Interaction,
    main_diff_dispos,
    int_diff_dispos,
    main_diff_disneg,
    int_diff_disneg), names_to = "predictor", values_to = "value")
  
  
boot_dat_mod_long_3 <- 
  boot_dat_mod_long_3 %>% 
  mutate(color = ifelse((predictor == "Disorganized_Main" & value > 0),
                        "diz",
                        ifelse((predictor == "Positive_Main" & value > 0),
                               "poz",
                        ifelse((predictor == "Negative_Main" & value > 0),
                               "neg",
                        ifelse((predictor == "main_diff_dispos" & value > 0),
                               "diz",
                        ifelse((predictor == "int_diff_dispos" & value > 0),
                               "diz",
                        ifelse((predictor == "main_diff_dispos" & value < 0),
                               "poz",
                        ifelse((predictor == "int_diff_dispos" & value < 0),
                               "poz",
                        ifelse((predictor == "main_diff_disneg" & value > 0),
                               "diz",
                        ifelse((predictor == "int_diff_disneg" & value > 0),
                               "diz",
                        ifelse((predictor == "main_diff_disneg" & value < 0),
                               "neg",
                        ifelse((predictor == "int_diff_disneg" & value < 0),
                               "neg",
                        ifelse((predictor == "Disorganized_Stress_Interaction" & value > 0),
                               "diz",
                        ifelse((predictor == "Positive_Stress_Interaction" & value > 0),
                               "poz",
                        ifelse((predictor == "Negative_Stress_Interaction" & value > 0),
                               "neg",
                               "graj"))))))))))))))) %>% 
  mutate(color = factor(color, levels = c("diz", "neg", "poz", "graj"), labels = c("#1B9E77", "#7570B3", "#d95f02","#6E6E6E")))


boot_dat_mod_9 <- 
  boot_dat_mod_9%>%
  select(-Stress_Main)
  
boot_dat_mod_long_9 <- 
  boot_dat_mod_9 %>%
    pivot_longer(c(
    Disorganized_Main, 
    Positive_Main,
    Negative_Main,
    Disorganized_Stress_Interaction,
    Positive_Stress_Interaction,
    Negative_Stress_Interaction,
    main_diff_dispos,
    int_diff_dispos,
    main_diff_disneg,
    int_diff_disneg), names_to = "predictor", values_to = "value")
  
  
boot_dat_mod_long_9 <- 
  boot_dat_mod_long_9 %>% 
  mutate(color = ifelse((predictor == "Disorganized_Main" & value > 0),
                        "diz",
                        ifelse((predictor == "Positive_Main" & value > 0),
                               "poz",
                        ifelse((predictor == "Negative_Main" & value > 0),
                               "neg",
                        ifelse((predictor == "main_diff_dispos" & value > 0),
                               "diz",
                        ifelse((predictor == "int_diff_dispos" & value > 0),
                               "diz",
                        ifelse((predictor == "main_diff_dispos" & value < 0),
                               "poz",
                        ifelse((predictor == "int_diff_dispos" & value < 0),
                               "poz",
                        ifelse((predictor == "main_diff_disneg" & value > 0),
                               "diz",
                        ifelse((predictor == "int_diff_disneg" & value > 0),
                               "diz",
                        ifelse((predictor == "main_diff_disneg" & value < 0),
                               "neg",
                        ifelse((predictor == "int_diff_disneg" & value < 0),
                               "neg",
                        ifelse((predictor == "Disorganized_Stress_Interaction" & value > 0),
                               "diz",
                        ifelse((predictor == "Positive_Stress_Interaction" & value > 0),
                               "poz",
                        ifelse((predictor == "Negative_Stress_Interaction" & value > 0),
                               "neg",
                               "graj"))))))))))))))) %>% 
  mutate(color = factor(color, levels = c("diz", "neg", "poz", "graj"), labels = c("#1B9E77", "#7570B3", "#d95f02","#6E6E6E")))
```

### differences between bootstrapped estimates of sch predictors (main and interaction effects)
##### PLEs

```{r}

boot_dat_mod_long_3 %>% 
  filter(predictor == "main_diff_disneg") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "neg")) %>%
  
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)


boot_dat_mod_long_3 %>% 
  filter(predictor == "int_diff_disneg") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "neg")) %>%
  
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)


boot_dat_mod_long_3 %>% 
  filter(predictor == "main_diff_dispos") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "pos")) %>%
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)

boot_dat_mod_long_3 %>% 
  filter(predictor == "int_diff_dispos") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "pos")) %>%
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)

```


```{r, fig.height = 4, fig.width = 8, warning=F}

m3_1 <- 
  boot_dat_mod_long_3 %>%
  filter(predictor == "main_diff_dispos") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.015) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
  labs(x = "Disorg. - Positive",
       y = "",
       title = "") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#d95f02"), c(0.7, 1)), 
                    limits = c("#1B9E77", "#d95f02"), 
                    labels =
                      c("Disorg.", "Positive")) +  
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  # scale_x_continuous(limits = c(-1, 1.5), breaks = seq(-1, 1, 0.5)) +
  scale_y_continuous(limits = c(0, 160), breaks = seq(0, 150, 50)) +
  annotate("text", x = c(-0.5,0.75), y = c(100, 100), label = c("", "81%"), size = 3) +
  theme(plot.margin = unit(c(0.3, 0, 0, 0), "cm")) +
  theme(plot.title = element_text(size = 10))+
  scale_x_continuous(labels = abs)

m3_2 <- 
  boot_dat_mod_long_3 %>%
  filter(predictor == "main_diff_disneg") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.015) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
   labs(x = "Disorg. - Negative",
       y = "",
       title = "Outcome: PLEs") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3"), c(0.7, 1)), 
                    limits = c("#1B9E77", "#7570B3"), 
                    labels =
                      c("Disorg.", "Negative")) +  
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  # scale_x_continuous(limits = c(-1, 1.5), breaks = seq(-1, 1, 0.5)) +
  scale_y_continuous(limits = c(0, 160), breaks = seq(0, 150, 50)) +
  annotate("text", x = c(-0.5,1), y = c(100, 100), label = c("", "100%"), size = 3) +
  theme(plot.margin = unit(c(0.3, 0, 0, 0), "cm"))+
  theme(plot.title = element_text(size = 10))+
  scale_x_continuous(labels = abs)

m3_3 <- 
  boot_dat_mod_long_3 %>%
  filter(predictor == "int_diff_dispos") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.05) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
  labs(x = "Disorg. - Positive",
       y = "",
       title = "") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#d95f02"), c(0.7, 1)),
                    limits = c("#1B9E77", "#d95f02"),
                    labels = c("Disorg.", "Positive")) +
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  scale_y_continuous(limits = c(0, 160), breaks = seq(0, 150, 50)) +
  # scale_x_continuous(limits = c(-1.5, 6.5), breaks = seq(-1, 6, 1)) +
  annotate("text", x = c(-1,5), y = c(100, 100), label = c("", "99%"), size = 3) +
  theme(plot.margin = unit(c(0.3, 0, 0, 1), "cm"))+
  theme(plot.title = element_text(size = 10)) +
  scale_x_continuous(labels = abs)


m3_4 <- 
  boot_dat_mod_long_3 %>%
  filter(predictor == "int_diff_disneg") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.05) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
  labs(x = "Disorg. - Negative",
       y = "",
       title = "Outcome: PLE-reactivity") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3"), c(0.7, 1)), 
                    limits = c("#1B9E77", "#7570B3"), 
                    labels =
                      c("Disorg.", "Negative")) +  
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  # scale_x_continuous(limits = c(-1.5, 6.5), breaks = seq(-1, 6, 1)) +
  scale_y_continuous(limits = c(0, 160), breaks = seq(0, 150, 50)) +
  annotate("text", x = c(-1,5), y = c(100, 100), label = c("", "100%"), size = 3) +
  theme(plot.margin = unit(c(0.3, 0, 0, 1), "cm")) +
  theme(plot.title = element_text(size = 10))+
  scale_x_continuous(labels = abs)

# (combined_plots_ple <- 
#   m3_1 + m3_3 +
#   m3_2 + m3_4 +
#   plot_layout(guides = "collect", ncol = 2, nrow = 2, heights = c(1.2, 1.2, 1.2)) + #, widths = c(1.2, 1.2, 1.2, 2)) +
# 
#   scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3", "#d95f02"), c(0.7, 1, 1)),
#                     limits = c("#1B9E77", "#7570B3", "#d95f02"),
#                     labels =
#                       c("Disorg.", "Negative", "Positive")) +
#   guides(fill = guide_legend(title = "")))
#     #    legend.position = 'right',
#     #    legend.direction = 'vertical')
```

### differences between bootstrapped estimates of sch predictors (main and interaction effects)
##### event-unpleasantness

```{r}

boot_dat_mod_long_9 %>% 
  filter(predictor == "main_diff_disneg") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "neg")) %>%
  
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)


boot_dat_mod_long_9 %>% 
  filter(predictor == "int_diff_disneg") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "neg")) %>%
  
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)


boot_dat_mod_long_9 %>% 
  filter(predictor == "main_diff_dispos") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "pos")) %>%
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)

boot_dat_mod_long_9 %>% 
  filter(predictor == "int_diff_dispos") %>%
  mutate(perc = ifelse(value > 0,
                       "diz",
                       "pos")) %>%
  select(perc) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(0)

```


```{r, fig.height = 4, fig.width = 8, warning=F}
m9_1 <- 
  boot_dat_mod_long_9 %>%
  filter(predictor == "main_diff_dispos") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.013) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
  labs(x = "Disorg. - Positive",
       y = "",
       title = "") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#d95f02"), c(0.7, 1)), 
                    limits = c("#1B9E77", "#d95f02"), 
                    labels =
                      c("Disorg.", "Positive")) +  
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  # scale_x_continuous(limits = c(-0.75, 1.5), breaks = seq(-0.5, 1, 0.5)) +
  scale_y_continuous(limits = c(0, 130), breaks = seq(0, 120, 40)) +
  annotate("text", x = c(-0.5,1), y = c(80, 80), label = c("", "99%"), size = 3) +
  theme(plot.title = element_text(size = 10)) +
  theme(plot.margin = unit(c(0.3, 0, 0, 1), "cm"))+
  scale_x_continuous(labels = abs)

m9_2 <- 
  boot_dat_mod_long_9 %>%
  filter(predictor == "main_diff_disneg") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.011) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
   labs(x = "Disorg. - Negative",
       y = "",
       title = "Outcome: Event-unpleasantness") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3"), c(0.7, 1)), 
                    limits = c("#1B9E77", "#7570B3"), 
                    labels =
                      c("Disorg.", "Negative")) +  
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  # scale_x_continuous(limits = c(-0.75, 1.5), breaks = seq(-0.5, 1, 0.5)) +
  scale_y_continuous(limits = c(0, 130), breaks = seq(0, 120, 40)) +
  annotate("text", x = c(-0.5, 0.5), y = c(80, 80), label = c("", "73%"), size = 3) +
  theme(plot.title = element_text(size = 10)) +
  theme(plot.margin = unit(c(0.3, 0, 0, 1), "cm"))+
  scale_x_continuous(labels = abs)

m9_3 <- 
  boot_dat_mod_long_9 %>%
  filter(predictor == "int_diff_dispos") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.03) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
  labs(x = "Disorg. - Positive",
       y = "",
       title = "") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#d95f02"), c(0.7, 1)),
                    limits = c("#1B9E77", "#d95f02"),
                    labels = c("Disorg.", "Positive")) +
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  # scale_x_continuous(limits = c(-1.5, 3.2), breaks = seq(-1, 3, 0.5)) +
  scale_y_continuous(limits = c(0, 130), breaks = seq(0, 120, 40)) +
  annotate("text", x = c(-0.5,2.5), y = c(80, 80), label = c("", "92%"), size = 3) +
  theme(plot.title = element_text(size = 10)) +
  theme(plot.margin = unit(c(0.3, 0, 0, 1), "cm"))+
  scale_x_continuous(labels = abs)

m9_4 <- 
  boot_dat_mod_long_9 %>%
  filter(predictor == "int_diff_disneg") %>%
  ggplot(aes(x = value)) +
  geom_histogram(aes(fill = color), binwidth = 0.03) + #, breaks = seq(-0.15, 0.15, by = 0.005)) +
  theme_minimal() +
  labs(x = "Disorg. - Negative",
       y = "",
       title = "Outcome: Stress-reactivity") +
  scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3"), c(0.7, 1)), 
                    limits = c("#1B9E77", "#7570B3"), 
                    labels =
                      c("Disorg.", "Negative")) +  
  guides(fill = "none") +
  geom_vline(xintercept = 0, color = "black", size = 0.7, alpha = 1) + 
  # scale_x_continuous(limits = c(-1.5, 3.2), breaks = seq(-1, 3, 0.5)) +
  scale_y_continuous(limits = c(0, 130), breaks = seq(0, 120, 40)) +
  annotate("text", x = c(-0.5,2.5), y = c(80, 80), label = c("", "100%"), size = 3) +
  theme(plot.title = element_text(size = 10)) +
  theme(plot.margin = unit(c(0.3, 0, 0, 1), "cm")) +
  scale_x_continuous(labels = abs)
```


```{r, warning = }

combined_plots_1 <- 
  m3_2 + m3_4 +
  m9_2 + m9_4 +
  m3_1 + m3_3 +
  m9_1 + m9_3 +
  plot_layout(guides = "collect",
              ncol = 4, nrow = 2) +
  scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3", "#d95f02"), c(0.7, 1, 1)),
                    limits = c("#1B9E77", "#7570B3", "#d95f02"),
                    labels = c("Disorg.", "Neg.", "Pos.")) +
  guides(fill = guide_legend(title = "")) +
  plot_annotation(title = "Bootstrapped estimates of the difference\nbetween model coefficients", theme = theme(plot.title = element_text(size = 12, face = "bold"))) & 
  theme(
   plot.title = element_text(hjust = 0.5),
   axis.title = element_text(size = 8, face = "bold"),
   axis.text = element_text(size = 8),
   legend.title = element_text(size = 8),
   legend.text = element_text(size = 8),
   legend.position = "bottom",
   legend.box = "horizontal")

ggsave("Fig04.jpeg", device = "jpeg", width = 28, height = 12, units = "cm")

combined_plots_1

```

```{r}

(boot_plot_1 <- 
  boot_dat_mod_long_3 %>% 
  filter(  predictor == "Disorganized_Stress_Interaction" | 
           predictor == "Positive_Stress_Interaction" | 
           predictor == "Negative_Stress_Interaction") %>%
  mutate(color = 
                        ifelse(predictor == "Disorganized_Stress_Interaction",
                               "diz",
                        ifelse(predictor == "Positive_Stress_Interaction",
                               "poz",
                               "neg"))) %>% 
  mutate(color = factor(color, levels = c("diz", "neg", "poz"), labels = c("#1B9E77", "#7570B3", "#d95f02"))) %>% 
  ggplot(aes(x = predictor, y = value, fill = color)) +
  geom_violin(color = NA, alpha = 0.5)+
  geom_boxplot(width = 0.2, fill = NA, outlier.shape = NA, lwd = 0.6) + 
  theme_classic() +
 # scale_fill_brewer(palette = 'Dark2') +
  guides(fill = 'none', color = 'none') +
  geom_hline(yintercept = 0, color = "black", linetype = 2, alpha = .4, size = .7) +
  labs(
       x  = "",
       y = "Bootstrapped interaction\nestimates",
       title = "",
       )  +
   scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3", "#d95f02"), 0.8), limits = c("#1B9E77", "#7570B3", "#d95f02"))  +
   scale_x_discrete(breaks = c("Disorganized_Stress_Interaction",
                               "Negative_Stress_Interaction",
                               "Positive_Stress_Interaction"),
         labels = c("Disorganized",
                    "Negative",
                    "Positive"))) +
  theme(axis.text.x = element_text(angle = -10, hjust = 0, vjust = 1))


(boot_plot_2 <- 
  boot_dat_mod_long_9 %>% 
  filter(  predictor == "Disorganized_Stress_Interaction" | 
           predictor == "Positive_Stress_Interaction" | 
           predictor == "Negative_Stress_Interaction") %>%
  mutate(color = 
                        ifelse(predictor == "Disorganized_Stress_Interaction",
                               "diz",
                        ifelse(predictor == "Positive_Stress_Interaction",
                               "poz",
                               "neg"))) %>% 
  mutate(color = factor(color, levels = c("diz", "neg", "poz"), labels = c("#1B9E77", "#7570B3", "#d95f02"))) %>% 
  ggplot(aes(x = predictor, y = value, fill = color)) +
  geom_violin(color = NA, alpha = 0.5)+
  geom_boxplot(width = 0.2, fill = NA, outlier.shape = NA, lwd = 0.6) + 
  theme_classic() +
 # scale_fill_brewer(palette = 'Dark2') +
  guides(fill = 'none', color = 'none') +
  geom_hline(yintercept = 0, color = "black", linetype = 2, alpha = .4, size = .7) +
  labs(
       x  = "",
       y = "",
       title = "",
       )  +
   scale_fill_manual(values = alpha(c("#1B9E77", "#7570B3", "#d95f02"), 0.8), limits = c("#1B9E77", "#7570B3", "#d95f02"))+
    scale_x_discrete(breaks = c("Disorganized_Stress_Interaction",
                                "Negative_Stress_Interaction",
                                "Positive_Stress_Interaction"),
          labels = c("Disorganized",
                     "Negative",
                     "Positive"))) +
  theme(axis.text.x = element_text(angle = -10, hjust = 0, vjust = 1))

```

```{r}
m3 <- readRDS("models/m3.rds")
m9 <- readRDS("models/m9.rds")


tab_model(m3, m9)
```


```{r}
neg_m3 <- plot_model(m3, type = "int")[[1]]
pos_m3 <- plot_model(m3, type = "int")[[2]]
dis_m3 <- plot_model(m3, type = "int")[[3]]

neg_m9 <- plot_model(m9, type = "int")[[1]]
pos_m9 <- plot_model(m9, type = "int")[[2]]
dis_m9 <- plot_model(m9, type = "int")[[3]]
```

```{r, warning=F}

dis_m3 <- 
  dis_m3 +
  geom_smooth(size = 1.3) +
  scale_fill_manual(labels = c("Low", "High"), values = c("black", "#1B9E77")) + 
  scale_colour_manual(labels = c("Low", "High"), values = c("black", "#1B9E77")) + 
  theme_classic() +
  theme(legend.position='none') +

  labs(
       x  = "",
       y = "",
       color = "Disorg.:",
       title = bquote("Prediction of" ~ italic("PLEs"))) 

neg_m3 <- 
  neg_m3 + 
  geom_smooth(size = 1.3) + 
  scale_fill_manual(labels = c("Low", "High"), values = c("black", "#7570B3")) + 
  scale_colour_manual(labels = c("Low", "High"), values = c("black", "#7570B3")) + 
  theme_classic() +
  theme(legend.position = 'none') +

  labs(
       x = "",
       y = "",
       color = "Negative:",
       title = "")

pos_m3_b <- 
  pos_m3 + 
  geom_smooth(size = 1.3) + 
  scale_fill_manual(labels = c("Low", "High"), values = c("black", "#d95f02")) + 
  scale_colour_manual(labels = c("Low", "High"), values = c("black", "#d95f02")) + 
  theme_classic() +
  theme(legend.position='none') +
  labs(
       x  = "Stress-exposure",
       y = "",
       color = "Positive:",
       title =  "") +
  theme(plot.title = element_text(hjust = 0.5))

dis_m9 <- 
  dis_m9 +
  geom_smooth(size = 1.3) + 
  scale_fill_manual(labels = c("Low", "High"), values = c("black", "#1B9E77")) + 
  scale_colour_manual(labels = c("Low", "High"), values = c("black", "#1B9E77")) + 
  theme_classic() +

  labs(
       x  = "",
       y = "",
       color = "Disorg.:",
       title = bquote("Prediction of" ~ italic("Event-unpleasantness")))

neg_m9 <- 
  neg_m9 + 
  geom_smooth(size = 1.3) + 
  scale_fill_manual(labels = c("Low", "High"), values = c("black", "#7570B3")) + 
  scale_colour_manual(labels = c("Low", "High"), values = c("black", "#7570B3")) + 
  theme_classic() +
  
  labs(
       x = "",
       y = "",
       color = "Negative:",
       title = "")

pos_m9_b <- 
  pos_m9 + 
  geom_smooth(size = 1.3) + 
  scale_fill_manual(labels = c("Low", "High"), values = c("black", "#d95f02")) + 
  scale_colour_manual(labels = c("Low", "High"), values = c("black", "#d95f02")) + 
  theme_classic() +
  labs(
       x  = "Stress-exposure",
       y = "",
       color = "Positive:",
       title =  "") +
  theme(plot.title = element_text(hjust = 0.5))

```

```{r, fig.height=7, fig.width=5, warning=F, message=F}

# add "fig.height=7, fig.width=5" to better visualization

combined_plots_2 <- 
  dis_m3 + dis_m9 +
  neg_m3 + neg_m9 + 
  pos_m3_b + pos_m9_b + 
  boot_plot_1 + 
  theme(axis.text.x = element_text(angle = -30, hjust = 0, vjust = 1)) +
 # scale_y_continuous(limits = c(-4, 4), breaks = seq(-4, 4, 2)) +
  boot_plot_2 + 
  theme(axis.text.x = element_text(angle = -30, hjust = 0, vjust = 1)) +
 # scale_y_continuous(limits = c(-4, 4), breaks = seq(-4, 4, 2)) +

  plot_layout(ncol = 2, nrow = 4, heights = c(1.2, 1.2, 1.2, 2))+ #, widths = c(1.2, 1.2, 1.2, 2)) + 
  plot_annotation(tag_levels = "A") & 
  theme(
    plot.title = element_text(size = 10, vjust = 6),
    plot.tag = element_text(size = 10, face = "bold"),
   # legend.position = "bottom",
    axis.title = element_text(size = 10),
    axis.text = element_text(size = 8),
    legend.title = element_text(size = 10),
    legend.text = element_text(size = 8))


combined_plots_2

ggsave("Fig03.jpeg", device = "jpeg", width = 15, height = 22, units = "cm")

```
