==> 04_descriptive_stats_ple_react.Rmd <==
---
title: "R Notebook"
author: "Flora Hann"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---


# loading packages
```{r}
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(readxl))
suppressPackageStartupMessages(library(lubridate))
suppressPackageStartupMessages(library(confintr))
suppressPackageStartupMessages(library(rstatix))
suppressPackageStartupMessages(library(ggrain))


options(scipen = 999)
```


# importing datasets
```{r message=FALSE}
initial_sample_pre <- read_csv("data_raw/dat_pre_cleaned.csv")
model_dat <- read_csv("data_processed/model_dat.csv")
dat_cleaned <- read_csv("data_processed/dat_cleaned.csv")
```

```{r}

model_dat <- 
  model_dat %>% 
  filter(!is.na(stressor_mean_centered_int) & !is.na(ple_cfa)) %>% 
  distinct()

```


# filtering datasets
```{r}
model_dat_distinct <- model_dat %>% 
  select(id_esm, session_esm, start_date, age, gender, edu) %>% 
  distinct() %>% 
  mutate(
    gender =
      if_else(
        gender == "Female",
        1,
        2
        ),
    gender = as.double(gender)
  )

dat_not_included <- initial_sample_pre %>% 
  anti_join(model_dat_distinct, by = c("session" = "session_esm", "gender", "edu")) %>% 
  filter(!is.na(session))
```


# sample demographics

## initial (=cross-sectional) sample
```{r}
# number and age of participants
  initial_sample_pre %>% 
    filter(!is.na(session)) %>% 
    select(session, age) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(age), digits = 2),
              Median    = round(median(age), digits = 2),
              SD        = round(sd(age), digits = 2),
              Range_min = min(age),
              Range_max = max(age))
```


```{r}
# gender
  # 1 = Female
  # 2 = Male

initial_sample_pre %>% 
  filter(!is.na(session)) %>% 
  select(session, gender) %>%
  distinct() %>% 
  select(gender) %>% 
  table()

initial_sample_pre %>% 
  filter(!is.na(session)) %>% 
  select(session, gender) %>%
  distinct() %>% 
  select(gender) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(2)
```


```{r}
# education
  # 1 = Primary school or lower
  # 3 = Vocational school without high school diploma
  # 4 = High school diploma or equivalent 
  # 5 = Bachelor's or Master's degree
  # 6 = Doctorate (PhD)
  # 7 = Other
  
initial_sample_pre %>% 
  filter(!is.na(session)) %>% 
  select(session, edu) %>%
  distinct() %>% 
  select(edu) %>% 
  table()

initial_sample_pre %>% 
  filter(!is.na(session)) %>% 
  select(session, edu) %>%
  distinct() %>% 
  select(edu) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(2)
```


```{r}
# MSS scores
  initial_sample_pre <- 
    initial_sample_pre %>% 
    # recode mss items 2="no"->0, 1="yes" stays the same
      mutate_at(
      vars(contains("MSS")),
      funs(
        ifelse(
          . == 2,
          0,
          .
        )
      )
    ) %>% 
    # revert reversed mss items 1->0, 0->1
    mutate(
      MSS_BH_4  = abs(MSS_BH_4  -1),
      MSS_BH_10 = abs(MSS_BH_10 -1),
      MSS_BH_25 = abs(MSS_BH_25 -1),
      MSS_BH_37 = abs(MSS_BH_37 -1)
      ) %>% 
    
    mutate(
      mss_pos = rowSums(select(., c("MSS_BH_2", "MSS_BH_5", "MSS_BH_8", "MSS_BH_11", "MSS_BH_14", "MSS_BH_17", "MSS_BH_20", "MSS_BH_23",   "MSS_BH_26", "MSS_BH_29", "MSS_BH_32", "MSS_BH_35", "MSS_BH_38"))),
      
      mss_neg = rowSums(select(., c("MSS_BH_1", "MSS_BH_4", "MSS_BH_7", "MSS_BH_10", "MSS_BH_13", "MSS_BH_16", "MSS_BH_19", "MSS_BH_22",   "MSS_BH_25", "MSS_BH_28", "MSS_BH_31", "MSS_BH_34", "MSS_BH_37"))),
               
      mss_dis = rowSums(select(., c("MSS_BH_3", "MSS_BH_6", "MSS_BH_9", "MSS_BH_12", "MSS_BH_15", "MSS_BH_18", "MSS_BH_21", "MSS_BH_24",   "MSS_BH_27", "MSS_BH_30", "MSS_BH_33", "MSS_BH_36")))
    )
  
  ## MSS_dis
    initial_sample_pre %>%
    select(session, age, gender, edu, mss_dis) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_dis), digits = 2),
              Median    = round(median(mss_dis), digits = 2),
              SD        = round(sd(mss_dis), digits = 2),
              Range_min = min(mss_dis),
              Range_max = max(mss_dis))
```


```{r}
## MSS_pos
    initial_sample_pre %>%
    select(session, age, gender, edu, mss_pos) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_pos), digits = 2),
              Median    = round(median(mss_pos), digits = 2),
              SD        = round(sd(mss_pos), digits = 2),
              Range_min = min(mss_pos),
              Range_max = max(mss_pos))
```


```{r}
## MSS_neg
    initial_sample_pre %>%
    select(session, age, gender, edu, mss_neg) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_neg), digits = 2),
              Median    = round(median(mss_neg), digits = 2),
              SD        = round(sd(mss_neg), digits = 2),
              Range_min = min(mss_neg),
              Range_max = max(mss_neg)) 
```


## not included
```{r}
# number and age of participants  
dat_not_included %>% 
  select(session, age) %>% 
  summarise(N         = n(),
            Mean      = round(mean(age), digits = 2),
            Median    = round(median(age), digits = 2),
            SD        = round(sd(age), digits = 2),
            Range_min = min(age),
            Range_max = max(age))
```


```{r}
# gender
  # 1 = Female
  # 2 = Male

dat_not_included %>% 
  select(gender) %>% 
  table()
  
dat_not_included %>%
  select(gender) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(2)
```


```{r}
# education
  # 1 = Primary school or lower
  # 3 = Vocational school without high school diploma
  # 4 = High school diploma or equivalent 
  # 5 = Bachelor's or Master's degree
  # 6 = Doctorate (PhD)
  # 7 = Other

dat_not_included %>% 
  select(edu) %>% 
  table()
  
dat_not_included %>% 
  select(edu) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(2)
```


```{r}
# MSS scores
  dat_not_included <- 
    dat_not_included %>% 
    # recode mss items 2="no"->0, 1="yes" stays the same
      mutate_at(
      vars(contains("MSS")),
      funs(
        ifelse(
          . == 2,
          0,
          .
        )
      )
    ) %>% 
    # revert reversed mss items 1->0, 0->1
    mutate(
      MSS_BH_4  = abs(MSS_BH_4  -1),
      MSS_BH_10 = abs(MSS_BH_10 -1),
      MSS_BH_25 = abs(MSS_BH_25 -1),
      MSS_BH_37 = abs(MSS_BH_37 -1)
      ) %>% 
    
    mutate(
      mss_pos = rowSums(select(., c("MSS_BH_2", "MSS_BH_5", "MSS_BH_8", "MSS_BH_11", "MSS_BH_14", "MSS_BH_17", "MSS_BH_20", "MSS_BH_23",   "MSS_BH_26", "MSS_BH_29", "MSS_BH_32", "MSS_BH_35", "MSS_BH_38"))),
      
      mss_neg = rowSums(select(., c("MSS_BH_1", "MSS_BH_4", "MSS_BH_7", "MSS_BH_10", "MSS_BH_13", "MSS_BH_16", "MSS_BH_19", "MSS_BH_22",   "MSS_BH_25", "MSS_BH_28", "MSS_BH_31", "MSS_BH_34", "MSS_BH_37"))),
               
      mss_dis = rowSums(select(., c("MSS_BH_3", "MSS_BH_6", "MSS_BH_9", "MSS_BH_12", "MSS_BH_15", "MSS_BH_18", "MSS_BH_21", "MSS_BH_24",   "MSS_BH_27", "MSS_BH_30", "MSS_BH_33", "MSS_BH_36")))
    )
  
  ## MSS_dis
    dat_not_included %>%
    select(session, age, gender, edu, mss_dis) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_dis), digits = 2),
              Median    = round(median(mss_dis), digits = 2),
              SD        = round(sd(mss_dis), digits = 2),
              Range_min = min(mss_dis),
              Range_max = max(mss_dis))
```


```{r}
## MSS_pos
    dat_not_included %>%
    select(session, age, gender, edu, mss_pos) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_pos), digits = 2),
              Median    = round(median(mss_pos), digits = 2),
              SD        = round(sd(mss_pos), digits = 2),
              Range_min = min(mss_pos),
              Range_max = max(mss_pos))
```


```{r}
## MSS_neg
    dat_not_included %>%
    select(session, age, gender, edu, mss_neg) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_neg), digits = 2),
              Median    = round(median(mss_neg), digits = 2),
              SD        = round(sd(mss_neg), digits = 2),
              Range_min = min(mss_neg),
              Range_max = max(mss_neg)) 
```


## final sample (proceeded to the ESM phase)
```{r}
# number and age of participants  
model_dat %>%
    select(id_esm, age) %>%
    distinct() %>%
    summarise(N         = n(),
              Mean      = round(mean(age), digits = 2),
              Median    = round(median(age), digits = 2),
              SD        = round(sd(age), digits = 2),
              Range_min = min(age),
              Range_max = max(age))
```


```{r}
# gender
model_dat %>% 
  select(id_esm, gender) %>%
  distinct() %>% 
  select(gender) %>% 
  table()

model_dat %>% 
  select(id_esm, gender) %>%
  distinct() %>% 
  select(gender) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(2)

  # 1 = Female
  # 2 = Male
```


```{r}
# education
model_dat %>% 
  select(id_esm, edu) %>%
  distinct() %>% 
  select(edu) %>% 
  table()

model_dat %>% 
  select(id_esm, edu) %>%
  distinct() %>% 
  select(edu) %>% 
  table() %>% 
  prop.table() %>% `*`(100) %>% round(2)
  
  # 1 = Primary school or lower
  # 3 = Vocational school without high school diploma
  # 4 = High school diploma or equivalent 
  # 5 = Bachelor's or Master's degree
  # 6 = Doctorate (PhD)
  # 7 = Other
```


```{r}
# MSS scores
  ## MSS_dis
    model_dat %>%
    select(id_esm, age, gender, edu, mss_dis) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_dis), digits = 2),
              Median    = round(median(mss_dis), digits = 2),
              SD        = round(sd(mss_dis), digits = 2),
              Range_min = min(mss_dis),
              Range_max = max(mss_dis))
```


```{r}
## MSS_pos
    model_dat %>%
    select(id_esm, age, gender, edu, mss_pos) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_pos), digits = 2),
              Median    = round(median(mss_pos), digits = 2),
              SD        = round(sd(mss_pos), digits = 2),
              Range_min = min(mss_pos),
              Range_max = max(mss_pos))
```


```{r}
## MSS_neg
    model_dat %>%
    select(id_esm, age, gender, edu, mss_neg) %>% 
    distinct() %>% 
    summarise(N         = n(),
              Mean      = round(mean(mss_neg), digits = 2),
              Median    = round(median(mss_neg), digits = 2),
              SD        = round(sd(mss_neg), digits = 2),
              Range_min = min(mss_neg),
              Range_max = max(mss_neg))
```


# descriptives

## descriptives of ESM surveys in final sample
```{r}
# total number of completed esm surveys
model_dat %>%
  distinct() %>% 
  summarise(n_id  = length(unique(id_esm)),
            n_obs = length(created_esm))
```


```{r}
#...and per capita
model_dat %>%
  filter(!is.na(created_esm)) %>%
  distinct() %>% 
  group_by(id_esm) %>% 
  summarise(n_obs = length(created_esm)) %>% 
  ungroup() %>% 
  summarise(max_n_obs    = max(n_obs),
            min_n_obs    = min(n_obs),
            mean_n_obs   = mean(n_obs),
            median_n_obs = median(n_obs),
            sd_n_obs     = sd(n_obs))
```


```{r}
# descriptives of ple
  model_dat %>%
    filter(!is.na(created_esm)) %>%
    select(id_esm, day, beep, ple_sum) %>%
    summarise(
      max_ple_sum    = max(ple_sum),
      min_ple_sum    = min(ple_sum),
      mean_ple_sum   = mean(ple_sum),
      median_ple_sum = median(ple_sum),
      sd_ple_sum    = sd(ple_sum)
  )
```


```{r}
# descriptives of event unpleasantness
  model_dat %>%
    filter(!is.na(created_esm)) %>%
    select(id_esm, day, beep, event_unpleasantness) %>%
    summarise(
      max_event_unpleasantness    = max(event_unpleasantness),
      min_event_unpleasantness    = min(event_unpleasantness),
      mean_event_unpleasantness   = mean(event_unpleasantness),
      median_event_unpleasantness = median(event_unpleasantness),
      sd_event_unpleasantness     = sd(event_unpleasantness)
  )
```


```{r}
# descriptives of stressor exposure
  model_dat %>% 
    filter(!is.na(date_diary) & !is.na(stressor_mean)) %>% 
    select(id_esm, date_diary, stressor_mean) %>%
    mutate(stressor_mean = stressor_mean * -1) %>% 
  
    unique() %>%
    summarise(
      max_stressor_mean    = max(stressor_mean),
      min_stressor_mean    = min(stressor_mean),
      mean_stressor_mean   = mean(stressor_mean),
      median_stressor_mean = median(stressor_mean),
      sd_stressor_mean     = sd(stressor_mean)
    )
```

## adding mss scores to model_dat_distinct
```{r}
model_dat_distinct <- model_dat %>% 
  select(id_esm, session_esm, start_date, age, gender, edu, mss_dis, mss_pos, mss_neg) %>% 
  distinct() %>% 
  mutate(
    gender =
      if_else(
        gender == "Female",
        1,
        2
        ),
    gender = as.double(gender)
  )
```


## statistical tests to compare initial and final samples
```{r}
# test for median age
  # normality test
  shapiro.test(dat_not_included$age)
  shapiro.test(model_dat_distinct$age)
    # -> normality not assumed -> mann-whitney
  
  # mann-whitney test
  wilcox.test(dat_not_included$age, model_dat_distinct$age)
  
  age_compare_not_included <- tibble(age = dat_not_included$age, group = "not_included")
  age_compare_final <- tibble(age = model_dat_distinct$age, group = "final")
  age_compare <- age_compare_not_included %>% bind_rows(age_compare_final)
  
  wilcox_test(age_compare, age ~ group, paired = FALSE) %>% add_significance()
  wilcox_effsize(age_compare, age ~ group, paired = FALSE)
```


```{r}
# test for gender distribution
  gender_compare_not_included <- tibble(gender = dat_not_included$gender, group = "not_included")
  gender_compare_final <- tibble(gender = model_dat_distinct$gender, group = "final")
  gender_compare <- gender_compare_not_included %>% bind_rows(gender_compare_final)
  
  chisq.test(gender_compare$group, gender_compare$gender)
```


```{r}
# test for education
  edu_compare_not_included <- tibble(edu = dat_not_included$edu, group = "not_included")
  edu_compare_final <- tibble(edu = model_dat_distinct$edu, group = "final")
  edu_compare <- edu_compare_not_included %>% bind_rows(edu_compare_final)
  
  edu_chisq <- chisq.test(edu_compare$group, edu_compare$edu, simulate.p.value = TRUE)
  
  edu_chisq
  cramersv(edu_chisq)
```


```{r}
# test for median mss scores
  
  ## mss dis
    # normality test
    shapiro.test(dat_not_included$mss_dis)
    shapiro.test(model_dat_distinct$mss_dis)
      # -> normality not assumed -> mann-whitney
  
    # mann-whitney test
    wilcox.test(dat_not_included$mss_dis, model_dat_distinct$mss_dis)
```


```{r}
## mss pos
    # normality test
    shapiro.test(dat_not_included$mss_pos)
    shapiro.test(model_dat_distinct$mss_pos)
      # -> normality not assumed -> mann-whitney
  
    # mann-whitney test
    wilcox.test(dat_not_included$mss_pos, model_dat_distinct$mss_pos)
```


```{r}
## mss neg
    # normality test
    shapiro.test(dat_not_included$mss_neg)
    shapiro.test(model_dat_distinct$mss_neg)
      # -> normality not assumed -> mann-whitney
  
    # mann-whitney test
    wilcox.test(dat_not_included$mss_neg, model_dat_distinct$mss_neg)
```


# calculating ESM survey compliance rate in final sample
```{r}
# identifying day of quitting
dat_maxday_esm <- model_dat %>% 
  select(id_esm, created_esm, day) %>% 
  filter(!is.na(created_esm)) %>% 
  group_by(id_esm) %>% 
  summarise(max_day_esm = max(day))

model_dat <- model_dat %>% 
  left_join(dat_maxday_esm, by = "id_esm")

# number of days spent in the study
model_dat %>% 
  select(id_esm, max_day_esm) %>%
  distinct() %>% 
  reframe(N               = n(),
          median_num_days = median(max_day_esm, na.rm = T),
          mean_num_days   = mean(max_day_esm, na.rm = T),
          sd_num_days     = sd(max_day_esm, na.rm = T),
          min_num_days    = min(max_day_esm, na.rm = T),
          max_num_days    = max(max_day_esm, na.rm = T)
  )

# calculating compliance rate (percentage)
model_dat %>%
  select(id_esm, created_esm, day, beep, max_day_esm) %>%
  distinct() %>%
  group_by(id_esm) %>%
  reframe(completed = sum(!is.na(created_esm)),
            all = max_day_esm * 8, 
            esm_compl_final = (completed / all) * 100,
  ) %>%
  ungroup() %>%
  reframe(n                  = n(),
          median_final_compl = median(esm_compl_final, na.rm = T),
          mean_final_compl   = mean(esm_compl_final, na.rm = T),
          sd_final_compl     = sd(esm_compl_final, na.rm = T),
          min_final_compl    = min(esm_compl_final, na.rm = T),
          max_final_compl    = max(esm_compl_final, na.rm = T)
  )

# calculating number of surveys completed
model_dat %>%
  select(id_esm, created_esm, day, beep, max_day_esm) %>%
  distinct() %>%
  group_by(id_esm) %>%
  reframe(completed = sum(!is.na(created_esm)),
          all = max_day_esm * 8, 
          esm_compl_final = (completed / all) * 100
  ) %>%
  ungroup() %>%
  reframe(n                    = n(),
          median_final_compl_n = median(completed, na.rm = T),
          mean_final_compl_n   = mean(completed, na.rm = T),
          sd_final_compl_n     = sd(completed, na.rm = T),
          min_final_compl_n    = min(completed, na.rm = T),
          max_final_compl_n    = max(completed, na.rm = T)
  )


# visualizing compliance rate
dat_figure_esm_compl <- model_dat %>%
  select(id_esm, created_esm, day, beep, max_day_esm) %>%
  distinct() %>% 
  group_by(id_esm) %>%
  reframe(completed = sum(!is.na(created_esm)),
          all = max_day_esm * 8,
          esm_compl_final = (completed / all) * 100
  ) %>%
  ungroup() %>%
  distinct()

visu_esm_compliance <- ggplot(dat_figure_esm_compl) +
  geom_histogram(aes(esm_compl_final), bins = 40, fill = "#A9A9A9") + #, color = "#797979") +
  theme_classic(base_size = 16) +
  labs(
    x = "Compliance rate (%)",
    y = "Number of participants"
  ) +
  theme(axis.title = element_text(size = 12))

visu_esm_compliance

# visualizing number of days
# ???
  
# correlation between ESM compliance and MSS-B subscores
## preprocessing
dat_corr_esm_compl_mssb <- model_dat %>% 
  select(id_esm, created_esm, day, beep, max_day_esm, mss_pos, mss_neg, mss_dis) %>%
  distinct() %>%
  group_by(id_esm) %>%
  mutate(completed = sum(!is.na(created_esm)),
          all = max_day_esm * 8, 
          esm_compl_final = (completed / all) * 100,
  ) %>% 
  select(id_esm, esm_compl_final, mss_pos, mss_neg, mss_dis) %>% 
  distinct()

## correlation for MSS-B positive subscale
### normality test
shapiro.test(dat_corr_esm_compl_mssb$esm_compl_final)
shapiro.test(dat_corr_esm_compl_mssb$mss_pos)
  # -> normality not assumed -> spearman

### Spearman correlation test
cor.test(dat_corr_esm_compl_mssb$esm_compl_final, dat_corr_esm_compl_mssb$mss_pos, method = "spearman")

## correlation for MSS-B negative subscale
### normality test
shapiro.test(dat_corr_esm_compl_mssb$mss_neg)
  # -> normality not assumed -> spearman

### Spearman correlation test
cor.test(dat_corr_esm_compl_mssb$esm_compl_final, dat_corr_esm_compl_mssb$mss_neg, method = "spearman")

## correlation for MSS-B disorganization subscale
### normality test
shapiro.test(dat_corr_esm_compl_mssb$mss_dis)
  # -> normality not assumed -> spearman

### Spearman correlation test
cor.test(dat_corr_esm_compl_mssb$esm_compl_final, dat_corr_esm_compl_mssb$mss_dis, method = "spearman")
```


# calculating three-day survey compliance rate in final sample
```{r}
# identifying day of quitting
dat_maxday_diary <- 
  model_dat %>%
  filter(!is.na(ended_diary)) %>%
  select(id_esm, three_day_block, ended_diary) %>%
  distinct() %>% 
  group_by(id_esm) %>%
  summarize(max_day_diary = max(three_day_block))

model_dat <- model_dat %>% 
  left_join(dat_maxday_diary, by = "id_esm")

# model_dat_missing <- model_dat %>% 
  # semi_join(missing, by = "id_esm")

# calculating compliance rate (percentage)
model_dat %>%
  select(id_esm, ended_diary, max_day_diary) %>%
  distinct() %>%
  group_by(id_esm) %>%
  reframe(completed = sum(!is.na(ended_diary)),
          all = max_day_diary, 
          diary_compl_final = (completed / all) * 100
  ) %>%
  ungroup() %>%
  reframe(median_final_compl_n = median(diary_compl_final, na.rm = T),
          mean_final_compl_n   = mean(diary_compl_final, na.rm = T),
          sd_final_compl_n     = sd(diary_compl_final, na.rm = T),
          min_final_compl_n    = min(diary_compl_final, na.rm = T),
          max_final_compl_n    = max(diary_compl_final, na.rm = T)
  )

# calculating number of surveys completed
model_dat %>%
  select(id_esm, ended_diary, max_day_diary) %>%
  distinct() %>%
  group_by(id_esm) %>%
  reframe(completed = sum(!is.na(ended_diary)),
          all = max_day_diary, 
          diary_compl_final = (completed / all) * 100
  ) %>%
  ungroup() %>%
  reframe(median_final_compl = median(completed, na.rm = T),
          mean_final_compl   = mean(completed, na.rm = T),
          sd_final_compl     = sd(completed, na.rm = T),
          min_final_compl    = min(completed, na.rm = T),
          max_final_compl    = max(completed, na.rm = T)
  )

# visualizing compliance rate
dat_figure_diary_compliance <- 
  model_dat %>%
  select(id_esm, ended_diary, max_day_diary) %>%
  distinct() %>%
  group_by(id_esm) %>%
  reframe(completed = sum(!is.na(ended_diary)),
          all = max_day_diary, 
          diary_compl_final = (completed / all) * 100
  ) %>%
  ungroup() %>%
  distinct()

visu_compliance_diary <- 
  ggplot(dat_figure_diary_compliance) +
  geom_histogram(aes(diary_compl_final), bins = 30,
                 fill = "#A9A9A9") + #, color = "#797979") +
  theme_classic() +
  labs(
    x = "Compliance rate (%)",
    y = "Number of participants"
  ) +
  theme(axis.title = element_text(size = 12))

visu_compliance_diary


# correlation between diary compliance and MSS-B subscores
## preprocessing
dat_corr_diary_compl_mssb <- model_dat %>% 
  select(id_esm, ended_diary, max_day_diary, mss_pos, mss_neg, mss_dis) %>%
  distinct() %>%
  group_by(id_esm) %>%
  mutate(completed = sum(!is.na(ended_diary)),
          all = max_day_diary, 
          diary_compl_final = (completed / all) * 100
  ) %>% 
  select(id_esm, diary_compl_final, mss_pos, mss_neg, mss_dis) %>% 
  distinct()

## correlation for MSS-B positive subscale
### normality test
shapiro.test(dat_corr_diary_compl_mssb$diary_compl_final)
shapiro.test(dat_corr_diary_compl_mssb$mss_pos)
  # -> normality not assumed -> spearman

### Spearman correlation test
cor.test(dat_corr_diary_compl_mssb$diary_compl_final, dat_corr_diary_compl_mssb$mss_pos, method = "spearman")

## correlation for MSS-B negative subscale
### normality test
shapiro.test(dat_corr_diary_compl_mssb$mss_neg)
  # -> normality not assumed -> spearman

### Spearman correlation test
cor.test(dat_corr_diary_compl_mssb$diary_compl_final, dat_corr_diary_compl_mssb$mss_neg, method = "spearman")

## correlation for MSS-B disorganization subscale
### normality test
shapiro.test(dat_corr_diary_compl_mssb$mss_dis)
  # -> normality not assumed -> spearman

### Spearman correlation test
cor.test(dat_corr_diary_compl_mssb$diary_compl_final, dat_corr_diary_compl_mssb$mss_dis, method = "spearman")
```

# distribution of PLE CFA scores
```{r}
ple_cfa_figure <- model_dat %>% 
  select(id_esm, day, beep, ple_cfa) %>%
  filter(!is.na(ple_cfa))

# histogram
hist_ple_cfa <- ggplot(ple_cfa_figure) +
  geom_histogram(aes(ple_cfa), bins = 30, fill = "#A9A9A9") +   #, color = "#797979") +
  theme_classic() +
  labs(
    x = "PLE factor score",
    y = "Number of cases"
  ) +
  scale_x_continuous(breaks=seq(-4,10,2)) +
  theme(axis.title = element_text(size = 12))

hist_ple_cfa
```

```{r, warning=F}

model_dat %>% 
  select(id_esm, mss_dis, mss_pos, mss_neg) %>% 
  distinct() %>% 
    
  gather("dimension", "value", 2:4) %>%
  
  ggplot(aes(dimension, value)) +
  
  ggdist::stat_halfeye(aes(fill = factor(dimension)), adjust = 1.5, width = .8, .width = 0, justification = -.18, point_colour = NA, position = position_dodge(width = 0.8), width = 0.25) + 
  geom_boxplot(width = .21, outlier.shape = NA, lwd = 0.6, position = position_dodge(width = 0.8), width = 0.25) + 
  geom_jitter(aes(color = factor(dimension)), width = .05, height = 0.1, alpha = .2, show.legend = F) +
 
  guides(fill = "none") + 
  xlab("") +
  ylab("Total raw scores") +
  labs(fill = "", title = "") + 
  scale_y_continuous(limits = c(0, 14), breaks = c(0, 2, 4, 6, 8, 10, 12, 14)) +
  theme_classic() +
  #+ 
  scale_fill_manual(labels = c("Disorganized", "Negative", "Positive"), values = alpha(c("#1B9E77", "#7570B3", "#d95f02"), 0.55)) +  
  
  scale_color_manual(labels = c("Disorganized", "Negative", "Positive"), values = c("#1B9E77", "#7570B3", "#d95f02")) + 
  scale_x_discrete(labels =c(
    "Disorganized",
    "Negative",
    "Positive"))  +
  theme(axis.text.x = element_text(angle = -30, hjust = 0, vjust = 1))

 # ggsave("distr.jpeg", device = "jpeg", width = 13, height = 9, units = "cm")

```

```{r}
# # raincloud plot
# raincloud_ple_cfa <-ggplot(ple_cfa_figure, aes(1, ple_cfa)) +
#   geom_rain(
#     alpha = .3
#     ) +
#   theme_classic() +
#   labs(
#     y = "PLE factor score",
#     x = ""
#   ) +
#   scale_y_continuous(breaks=seq(-4,10,2)) +
#   theme(
#     axis.title = element_text(size = 12),
#     axis.ticks.y = element_blank(),
#     axis.text.y = element_blank()
#     ) +
#   coord_flip()
# 
# raincloud_ple_cfa
```
