==> 05_power_analysis.Rmd <==
---
title: "05_power_analysis"
author: "Levente Ronai"
date: "2024-04-08"
output: html_document
---

```{r, include=F}

options(scipen = 999)

suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(dplyr))


```


```{r}

model_dat <- read.csv("data_processed/model_dat.csv")

m3 <- readRDS("models/m3.rds")
m9 <- readRDS("models/m9.rds")

```

```{r}
summary(m3)
```

```{r, include=F}
# Check if R packages are installed

# list.of.packages = c("htmltools","shiny","htmltools","shiny","DT","nlme","ggplot2","gridExtra",
# "data.table","plyr","dplyr","formattable","tidyr","MASS","shinyjs","compiler","future.apply","devtools")
# new.packages = list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
# if(length(new.packages)) install.packages(new.packages)

library(htmltools)
library(shiny)
library(DT)
library(nlme)
library(ggplot2)
library(gridExtra)
library(data.table)
library(plyr)
library(dplyr)
library(formattable)
library(tidyr)
library(MASS)
library(shinyjs)
library(compiler)
library(future.apply)

# library(devtools)
# devtools::install_github("ginettelafit/PowerAnalysisIL", force = T)

library(PowerAnalysisIL)



# using the following parameters:

# model 7, cross level interaction with random slope
# N participants: 100,126
# N obs: 4611
# Fixed intercept: -0.0548185
# Effect of the level-2 continuous variable on the intercept: 0.4391183
# Fixed slope: 0.3745155
# Effect of the level-2 continuous variable on the slope: 1.8491296
# Standard deviation of level-1 errors: 0.7790
# Autocorrelation of level-1 errors: ~ 0.2
# Standard deviation of random intercept: 0.8678
# Standard deviation of random slope: 2.3483
# Correlation between the random intercept and random slope: 0.06
# Mean of time-varying variable X: 0
# Standard deviation of time-varying variable X: 0.13
# without person-mean centering
# Mean of level-2 variable W: 0
# Standard deviation of level-2 variable W: 1
# without centering and estimating autocorrelation
# type I: 0,05
# replicates: 200
# method: max. loglikelihood.
# 1, estimate comp. time; 2, comp. power

# Using Gist: users can launch this app with:
shiny::runGist('6bac9d35c2521cc4fd91ce4b82490236')


```
